@@ -6,8 +6,8 @@
   "license": "LicenseRef-LICENSE",
   "version": "0.0.1",
   "engines": {
-    "node": "0.12.2",
-    "npm": "2.8.3"
+    "node": ">=4.0.0",
+    "npm": ">=3.0.0"
   },
   "repository": {
     "type": "git",
@@ -25,22 +25,25 @@
     "bytebuffer": "^5.0.0",
     "crypto-js": "3.1.2-3",
     "ecurve": "1.0.0",
+    "event-emitter": "^0.3.4",
     "immutable": "^3.7.5",
     "lodash": "^3.10.1",
     "lzma": "2.1.6",
     "numeral": "^1.5.3",
     "qr-image": "3.1.0",
-    "react": "^0.13.3",
+    "react": "^0.14.4",
+    "react-dom": "^0.14.4",
     "rx": "^2.5.2",
     "secure-random": "1.1.1",
     "tcomb": "2.5.2",
     "websocket": "^1.0.18"
   },
   "devDependencies": {
-    "babel": "^5.6.14",
-    "babel-core": "~5.8.33",
-    "babel-loader": "~5.3.1",
+    "babel": "^5.8.29",
+    "babel-core": "^5.8.33",
+    "babel-loader": "^5.3.3",
     "babel-runtime": "^5.6.15",
+    "babel-plugin-react-transform": "^1.1.1",
     "coffee-loader": "~0.7.2",
     "coffee-script": "~1.8.0",
     "coffeeify": "~0.7.0",
@@ -50,9 +53,6 @@
     "node-fetch": "^1.3.1",
     "node-libs-browser": "~0.5.2",
     "node-localstorage": "~0.3.4",
-    "react-tools": "~0.10.0",
-    "react-transform": "0.0.3",
-    "traceur": "0.0.92",
-    "webpack": "~1.10.1"
+    "traceur": "0.0.92"
   }
 }
