@@ -6,8 +6,8 @@
   "author": "Cryptonomex, Inc.",
   "license": "LicenseRef-LICENSE",
   "engines": {
-    "node": "0.12.2",
-    "npm": "2.8.3"
+    "node": ">=4.0.0",
+    "npm": ">=3.0.0"
   },
   "repository": {
     "type": "git",
@@ -23,7 +23,6 @@
   "dependencies": {
     "alt": "^0.17.8",
     "alt-container": "^1.0.0",
-    "babel": "^5.8.29",
     "bignumber.js": "^2.1.0",
     "classnames": "^2.2.1",
     "cookies-js": "^1.2.1",
@@ -41,20 +40,20 @@
     "lodash": "^3.10.1",
     "node-fetch": "^1.3.1",
     "object-assign": "^4.0.1",
-    "perfect-scrollbar": "^0.6.7",
+    "perfect-scrollbar": "^0.6.10",
     "react": "^0.14.4",
+    "react-addons-css-transition-group": "^0.14.7",
     "react-dom": "^0.14.4",
     "react-foundation-apps": "git+http://github.com/valzav/react-foundation-apps.git",
     "react-highcharts": "^6.0.0",
     "react-interpolate-component": "^0.8.0",
-    "react-intl": "^2.0.0-beta-2",
+    "react-intl": "2.0.0-beta-2",
     "react-json-inspector": "6.0.0",
     "react-motion": "0.3.1",
     "react-notification-system": "^0.2.4",
-    "react-popover": "^0.4.0",
+    "react-popover": "^0.4.4",
     "react-responsive-mixin": "^0.3.5",
     "react-router": "1.0.1",
-    "react-tools": "~0.13.2",
     "react-tooltip": "^0.9.3",
     "react-translate-component": "^0.10.0",
     "remarkable-loader": "^0.1.0",
@@ -64,6 +63,7 @@
   },
   "devDependencies": {
     "autoprefixer-loader": "^3.1.0",
+    "babel": "^5.8.29",
     "babel-core": "^5.8.33",
     "babel-eslint": "^4.1.4",
     "babel-jest": "^5.3.0",
