@@ -5,6 +5,7 @@ import Highcharts from "react-highcharts/bundle/highstock";
 import utils from "common/utils";
 import _ from "lodash";
 import Translate from "react-translate-component";
+import colors from "assets/colors";
 
 require("./highcharts-plugins/technical-indicators.src.js");
 require("./highcharts-plugins/rsi.js");
@@ -83,7 +84,10 @@ class PriceChart extends React.Component {
             nextProps.baseSymbol !== this.props.baseSymbol ||
             latestCheck ||
             nextProps.leftOrderBook !== this.props.leftOrderBook ||
-            !utils.are_equal_shallow(nextProps.indicatorSettings, this.props.indicatorSettings)
+            !utils.are_equal_shallow(nextProps.indicatorSettings, this.props.indicatorSettings) ||
+            nextProps.verticalOrderbook !== this.props.verticalOrderbook ||
+            nextProps.height !== this.props.height ||
+            nextProps.zoom !== this.props.zoom
         );
     }
 
@@ -97,9 +101,28 @@ class PriceChart extends React.Component {
         };
     }
 
-    componentWillReceiveProps() {
+    componentDidMount() {
+        this.reflowChart(500);
+    }
+
+    componentWillReceiveProps(nextProps) {
         let height = ReactDOM.findDOMNode(this).offsetHeight;
         this.setState({offsetHeight: height - 10});
+
+        if (this.refs.chart && 
+            (nextProps.verticalOrderbook !== this.props.verticalOrderbook ||
+            nextProps.height !== this.props.height)
+                ) {
+            this.reflowChart(100);
+        }
+    }
+
+    reflowChart(timeout) {
+        setTimeout(() => {
+            if (this.refs.chart) {
+                this.refs.chart.chart.reflow();
+            }
+        }, timeout);   
     }
 
     getIndicators(props, select = false) {
@@ -200,13 +223,13 @@ class PriceChart extends React.Component {
 
     render() {
         let {priceData, volumeData, quoteSymbol, baseSymbol, base, quote, marketReady,
-            indicators, indicatorSettings, latest, bucketSize} = this.props;
+            indicators, indicatorSettings, latest, bucketSize, theme} = this.props;
 
         let priceSeriesData = _.cloneDeep(priceData);
         let currentIndicator = this.getIndicators(this.props);
 
-        let positiveColor = "rgba(110, 193, 5, 0.80)";
-        let negativeColor = "rgba(225, 66, 74, 0.80)";
+        let positiveColor = colors[theme].positiveColor;
+        let negativeColor = colors[theme].negativeColor;
         
         if (!priceSeriesData.length && latest) {
             let now = (new Date).getTime();
@@ -221,18 +244,27 @@ class PriceChart extends React.Component {
             negativeColor = "black";
         }
 
-        let maxVolume = 0;
+        // Find max volume
+        // let maxVolume = 0;
         let volumeColors = [], colorByPoint = false;
 
-        if (volumeData.length === priceSeriesData.length) {
-            colorByPoint = true;
-        }
-        for (var i = 0; i < volumeData.length; i++) {
-            maxVolume = Math.max(maxVolume, volumeData[i][1]);
-            if (colorByPoint) {
-                volumeColors.push(priceSeriesData[i][1] <= priceSeriesData[i][4] ? positiveColor : negativeColor);
-            }
-        }
+        // if (volumeData.length === priceSeriesData.length) {
+        //     colorByPoint = true;
+        // }
+        // for (var i = 0; i < volumeData.length; i++) {
+        //     maxVolume = Math.max(maxVolume, volumeData[i][1]);
+            // if (colorByPoint) {
+            //     volumeColors.push(priceSeriesData[i][1] <= priceSeriesData[i][4] ? positiveColor : negativeColor);
+            // }
+        // }
+
+        // Find highest price
+        // let maxPrice = 0;
+        // if (priceSeriesData.length) {
+        //     for (var i = 0; i < priceSeriesData.length; i++) {
+        //         maxPrice = Math.max(maxPrice, priceSeriesData[i][2]);
+        //     }
+        // }
 
         let config = {
             chart: {
@@ -241,7 +273,8 @@ class PriceChart extends React.Component {
                     enabled: false
                 },
                 pinchType: "x",
-                spacing: [20, 10, 5, 10]
+                spacing: [20, 10, 5, 10],
+                alignTicks: false
             },
 
             indicators: priceSeriesData.length ? currentIndicator : [],
@@ -258,7 +291,9 @@ class PriceChart extends React.Component {
                 enabled: false
             },
             navigator: {
-                enabled: true
+                enabled: true,
+                height: 30,
+                margin: 10
             },
             rangeSelector: {
                 enabled: false
@@ -300,16 +335,16 @@ class PriceChart extends React.Component {
                     if (!this.points || this.points.length === 0) {
                         return "";
                     }
-                    let TA = _(this.points[0].indicators).reduce((finalString, indicator, key) => {
+                    let TA = _(this.points[1].indicators).reduce((finalString, indicator, key) => {
                         return finalString + "<b>" + key.toUpperCase() + "</b>" + ": " + Highcharts.Highcharts.numberFormat(indicator[1], price_dec, ".", ",") + "  ";
                     }, "");
 
                     return ("<span style='color: white;fill: white'><b>T:&nbsp;</b>" + time +
-                            "&nbsp;<b>O:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[0].point.open, price_dec, ".", ",") +
-                            "&nbsp;&nbsp;<b>H:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[0].point.high, price_dec, ".", ",") +
-                            "&nbsp;&nbsp;<b>L:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[0].point.low, price_dec, ".", ",") +
-                            "&nbsp;&nbsp;<b>C:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[0].point.close, price_dec, ".", ",") +
-                            "<b>&nbsp;V:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[1] ? this.points[1].point.y : 0, vol_dec, ".", ",") + " " +
+                            "&nbsp;<b>O:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[1].point.open, price_dec, ".", ",") +
+                            "&nbsp;&nbsp;<b>H:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[1].point.high, price_dec, ".", ",") +
+                            "&nbsp;&nbsp;<b>L:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[1].point.low, price_dec, ".", ",") +
+                            "&nbsp;&nbsp;<b>C:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[1].point.close, price_dec, ".", ",") +
+                            "<b>&nbsp;V:&nbsp;</b>" + Highcharts.Highcharts.numberFormat(this.points[1] ? this.points[0].point.y : 0, vol_dec, ".", ",") + " " +
                             quoteSymbol + "<br/>" + TA + "</span>");
 
                 },
@@ -319,19 +354,18 @@ class PriceChart extends React.Component {
             },
             series: [
                 {
+                    type: "column",
+                    name: `Volume`,
+                    data: volumeData,
+                    color: colors[theme].volumeColor,
+                    yAxis: 1
+                },
+                {
                     id: "primary",
                     type: "candlestick",
                     name: `Price`,
                     data: priceSeriesData
-                },
-                {
-                    type: "column",
-                    name: `Volume`,
-                    data: volumeData,
-                    color: "#E3745B",
-                    yAxis: 1
                 }
-
             ],
             yAxis: [{
                     labels: {
@@ -349,29 +383,29 @@ class PriceChart extends React.Component {
                             color: "#FFFFFF"
                         }
                     },
-                    top: "0%",
-                    height: "80%",
                     offset: 5,
+                    lineWidth: 1,
+                    lineColor: "rgba(183, 183, 183, 0.29)",
                     gridLineWidth: 0,
                     plotLines: [],
                     crosshair: {
                         snap: false
                     },
                     startOnTick: false,
-                    endOnTick: false,
+                    endOnTick: true,
                     showLastLabel: true,
                     maxPadding: 0,
                     currentPriceIndicator: {
                         precision: base.get("precision"),
-                        backgroundColor: '#000000',
+                        backgroundColor: '#C38B8B',
                         borderColor: '#000000',
-                        lineColor: '#000000',
+                        lineColor: '#C38B8B',
                         lineDashStyle: 'Solid',
-                        lineOpacity: 0.6,
+                        lineOpacity: 0.8,
                         enabled: priceSeriesData.length > 0 && marketReady,
                         style: {
                             color: '#ffffff',
-                            fontSize: '12px'
+                            fontSize: '10px'
                         },
                         x: -30,
                         y: 0,
@@ -400,11 +434,13 @@ class PriceChart extends React.Component {
                             }
                         }
                     },
-                    opposite: true,
-                    top: "80%",
-                    height: "20%",
+                    opposite: false,
                     offset: 5,
                     gridLineWidth: 0,
+                    lineWidth: 1,
+                    lineColor: "rgba(183, 183, 183, 0.29)",
+                    endOnTick: true,
+                    showLastLabel: true,
                     title: {
                         text: null,
                         style: {
@@ -412,9 +448,7 @@ class PriceChart extends React.Component {
                         }
                     },
                     showFirstLabel: true,
-                    tickInterval: Math.floor(maxVolume / 2.5),
-                    min: 0,
-                    max: maxVolume
+                    min: 0
             }],
             xAxis: {
                 type: "datetime",
@@ -428,7 +462,8 @@ class PriceChart extends React.Component {
                 title: {
                     text: null
                 },
-                plotLines: []
+                plotLines: [],
+                min: this.props.zoom === "all" ? null : new Date().getTime() - 1000 * this.props.zoom
 
             }
         };
@@ -467,13 +502,20 @@ class PriceChart extends React.Component {
 
         let boxHeight = 20;
 
-
-
         return (
-            <div className="grid-content no-padding no-overflow">
-                {!priceSeriesData.length ? <span className="no-data"><Translate content="exchange.no_data" /></span> : null}
-                <div style={{paddingTop: 0, paddingBottom: "0.5rem"}}>
-                    {priceSeriesData && volumeData ? <Highcharts ref="chart" config={config}/> : null}
+            <div className="grid-content no-padding no-margin no-overflow middle-content">
+                <div className="exchange-bordered" style={{margin: 10}}>
+                    <div className="exchange-content-header">
+                        <Translate content="exchange.price_history" />
+                        <div className="float-right">
+                            <div style={{display: "inline-block", marginBottom: -3, marginTop: -6, padding: "3px 8px"}} className="button outline clickable" onClick={this.props.onChangeSize.bind(this, false)}>-</div>
+                            <div style={{display: "inline-block", marginBottom: -3, marginTop: -6, padding: "3px 8px"}} className="button outline clickable" onClick={this.props.onChangeSize.bind(this, true)}>+</div>
+                        </div>
+                    </div>
+                    {!priceSeriesData.length ? <span className="no-data"><Translate content="exchange.no_data" /></span> : null}
+                    <div style={{paddingTop: 0, paddingBottom: "0.5rem"}}>
+                        {priceSeriesData && volumeData ? <Highcharts ref="chart" config={config}/> : null}
+                    </div>
                 </div>
             </div>
         );
