@@ -31,10 +31,15 @@ class MarketUtils {
         return op.amount_to_sell.asset_id !== op.fee.asset_id;
     }
 
-    static limitByPrecision(value, asset) {
-
-        let precision = utils.get_asset_precision(asset.toJS ? asset.get("precision") : asset.precision);
-        value = Math.floor(value * precision) / precision;
+    static limitByPrecision(value, asset, floor = true) {
+        let assetPrecision = asset.toJS ? asset.get("precision") : asset.precision;
+        let valueString = value.toString();
+        let splitString = valueString.split(".");
+        if (splitString.length === 1 || splitString.length === 2 && splitString[1].length <= assetPrecision) {
+            return value;
+        }
+        let precision = utils.get_asset_precision(assetPrecision);
+        value = floor ? Math.floor(value * precision) / precision : Math.round(value * precision) / precision;
         if (isNaN(value) || !isFinite(value)) {
             return 0;
         }
@@ -89,7 +94,7 @@ class MarketUtils {
             buy.amount = parseInt(buy.amount, 10);
         }
         let fullPrice = callPrice ? callPrice : (sell.amount / basePrecision) / (buy.amount / quotePrecision)
-        let price = utils.price_to_text(fullPrice, quote, base);
+        let price = utils.price_to_text(fullPrice, order.call_price ? base : quote, order.call_price ? quote : base);
 
         let amount, value;
 
@@ -140,9 +145,9 @@ class MarketUtils {
         let payPrecision = utils.get_asset_precision(paysAsset.get("precision"));
 
         let receives = order.receives.amount / receivePrecision;
-        receives = utils.format_number(receives, receivesAsset.get("precision") - 1);
+        receives = utils.format_number(receives, receivesAsset.get("precision"));
         let pays = order.pays.amount / payPrecision;
-        pays = utils.format_number(pays, paysAsset.get("precision") - 1);
+        pays = utils.format_number(pays, paysAsset.get("precision"));
         let price_full = utils.get_asset_price(order.receives.amount, receivesAsset, order.pays.amount, paysAsset, isAsk);
         // price_full = !flipped ? (1 / price_full) : price_full;
         // let {int, dec} = this.split_price(price_full, isAsk ? receivesAsset.get("precision") : paysAsset.get("precision"));
@@ -190,89 +195,89 @@ class MarketUtils {
         return {int: int, dec: dec};
     }
 
-    static flatten_orderbookchart(array, sumBoolean, inverse, precision) {
-        inverse = inverse === undefined ? false : inverse;
-        let orderBookArray = [];
-        let maxStep, arrayLength = array.length;
+    // static flatten_orderbookchart(array, sumBoolean, inverse, precision) {
+    //     inverse = inverse === undefined ? false : inverse;
+    //     let orderBookArray = [];
+    //     let maxStep, arrayLength = array.length;
 
-        // Sum orders at same price
-        // if (arrayLength > 1) {
-        //     for (var i = arrayLength - 2; i >= 0; i--) {
-        //         if (array[i].x === array[i + 1].x) {
-        //             console.log("found order to sum");
-        //             array[i].y += array[i + 1].y;
-        //             array.splice(i + 1, 1);
-        //         }
-        //     }
-        // }
-        // arrayLength = array.length;
+    //     // Sum orders at same price
+    //     // if (arrayLength > 1) {
+    //     //     for (var i = arrayLength - 2; i >= 0; i--) {
+    //     //         if (array[i].x === array[i + 1].x) {
+    //     //             console.log("found order to sum");
+    //     //             array[i].y += array[i + 1].y;
+    //     //             array.splice(i + 1, 1);
+    //     //         }
+    //     //     }
+    //     // }
+    //     // arrayLength = array.length;
 
-        if (inverse) {
+    //     if (inverse) {
 
-            if (array && arrayLength) {
-                arrayLength = arrayLength - 1;
-                orderBookArray.unshift({
-                    x: array[arrayLength].x,
-                    y: array[arrayLength].y
-                });
-                if (array.length > 1) {
-                    for (let i = array.length - 2; i >= 0; i--) {
-                        maxStep = Math.min((array[i + 1].x - array[i].x) / 2, 0.1 / precision);
-                        orderBookArray.unshift({
-                            x: array[i].x + maxStep,
-                            y: array[i + 1].y
-                        });
-                        if (sumBoolean) {
-                            array[i].y += array[i + 1].y;
-                        }
-                        orderBookArray.unshift({
-                            x: array[i].x,
-                            y: array[i].y
-                        });
-                    }
-                } else {
-                    orderBookArray.unshift({
-                        x: 0,
-                        y: array[arrayLength].y
-                    });
-                }
-            }
-        } else {
-            if (array && arrayLength) {
-                orderBookArray.push({
-                    x: array[0].x,
-                    y: array[0].y
-                });
-                if (array.length > 1) {
-                    for (let i = 1; i < array.length; i++) {
-                        maxStep = Math.min((array[i].x - array[i - 1].x) / 2, 0.1 / precision);
-                        orderBookArray.push({
-                            x: array[i].x - maxStep,
-                            y: array[i - 1].y
-                        });
-                        if (sumBoolean) {
-                            array[i].y += array[i - 1].y;
-                        }
-                        orderBookArray.push({
-                            x: array[i].x,
-                            y: array[i].y
-                        });
-                    }
-                } else {
-                    orderBookArray.push({
-                        x: array[0].x * 1.5,
-                        y: array[0].y
-                    });
-                }
-            }
-        }
-        return orderBookArray;
-    }
+    //         if (array && arrayLength) {
+    //             arrayLength = arrayLength - 1;
+    //             orderBookArray.unshift({
+    //                 x: array[arrayLength].x,
+    //                 y: array[arrayLength].y
+    //             });
+    //             if (array.length > 1) {
+    //                 for (let i = array.length - 2; i >= 0; i--) {
+    //                     // maxStep = Math.min((array[i + 1].x - array[i].x) / 2, 0.1 / precision);
+    //                     orderBookArray.unshift({
+    //                         x: array[i].x + maxStep,
+    //                         y: array[i + 1].y
+    //                     });
+    //                     if (sumBoolean) {
+    //                         array[i].y += array[i + 1].y;
+    //                     }
+    //                     orderBookArray.unshift({
+    //                         x: array[i].x,
+    //                         y: array[i].y
+    //                     });
+    //                 }
+    //             } else {
+    //                 orderBookArray.unshift({
+    //                     x: 0,
+    //                     y: array[arrayLength].y
+    //                 });
+    //             }
+    //         }
+    //     } else {
+    //         if (array && arrayLength) {
+    //             orderBookArray.push({
+    //                 x: array[0].x,
+    //                 y: array[0].y
+    //             });
+    //             if (array.length > 1) {
+    //                 for (let i = 1; i < array.length; i++) {
+    //                     // maxStep = Math.min((array[i].x - array[i - 1].x) / 2, 0.1 / precision);
+    //                     orderBookArray.push({
+    //                         x: array[i].x - maxStep,
+    //                         y: array[i - 1].y
+    //                     });
+    //                     if (sumBoolean) {
+    //                         array[i].y += array[i - 1].y;
+    //                     }
+    //                     orderBookArray.push({
+    //                         x: array[i].x,
+    //                         y: array[i].y
+    //                     });
+    //                 }
+    //             } else {
+    //                 orderBookArray.push({
+    //                     x: array[0].x * 1.5,
+    //                     y: array[0].y
+    //                 });
+    //             }
+    //         }
+    //     }
+    //     return orderBookArray;
+    // }
 
     static flatten_orderbookchart_highcharts(array, sumBoolean, inverse, precision) {
         inverse = inverse === undefined ? false : inverse;
         let orderBookArray = [];
-        let maxStep, arrayLength;
+        let arrayLength;
 
         if (inverse) {
 
@@ -281,8 +286,6 @@ class MarketUtils {
                 orderBookArray.unshift([array[arrayLength][0], array[arrayLength][1]]);
                 if (array.length > 1) {
                     for (let i = array.length - 2; i >= 0; i--) {
-                        maxStep = Math.min((array[i + 1][0] - array[i][0]) / 2, 0.1 / precision);
-                        orderBookArray.unshift([array[i][0] + maxStep, array[i + 1][1]]);
                         if (sumBoolean) {
                             array[i][1] += array[i + 1][1];
                         }
@@ -297,8 +300,6 @@ class MarketUtils {
                 orderBookArray.push([array[0][0], array[0][1]]);
                 if (array.length > 1) {
                     for (var i = 1; i < array.length; i++) {
-                        maxStep = Math.min((array[i][0] - array[i - 1][0]) / 2, 0.1 / precision);
-                        orderBookArray.push([array[i][0] - maxStep, array[i - 1][1]]);
                         if (sumBoolean) {
                             array[i][1] += array[i - 1][1];
                         }
@@ -331,6 +332,25 @@ class MarketUtils {
         } else {
             throw "Unknown type";
         }
+    }
+
+    static isMarketAsset(quote, base) {
+        let isMarketAsset = false, marketAsset, inverted = false;
+
+        if (quote.get("bitasset") && base.get("id") === "1.3.0") {
+            isMarketAsset = true;
+            marketAsset = {id: quote.get("id")}
+        } else if (base.get("bitasset") && quote.get("id") === "1.3.0") {
+            inverted = true;
+            isMarketAsset = true;
+            marketAsset = {id: base.get("id")};
+        }
+
+        return {
+            isMarketAsset,
+            marketAsset,
+            inverted
+        };
     }
 
 }
