@@ -15,7 +15,7 @@
         dashboard: "Dashboard",
         explorer: "Explore",
         exchange: "Trade",
-        payments: "Transfer", // temporary, once we have tabs on Payments, I'll change it back to payments (Valentine)
+        payments: "Send", // temporary, once we have tabs on Payments, I'll change it back to payments (Valentine)
         logout: "Logout",
         settings: "Settings",
         current: "Current Account",
@@ -106,7 +106,9 @@
             required: "Coin days required",
             remaining: "Days left of vesting period",
             claim: "Claim now",
-            available: "Available to claim"
+            available: "Available to claim",
+            referral_link: "Your referral link",
+            referral_text: "Give this to link to people you want to refer to Bitshares"
         },
         user_issued_assets: {
             symbol: "Symbol",
@@ -144,7 +146,23 @@
             advanced: "Advanced",
             need_asset: "At least one of the two assets must be %(name)s",
             perm_warning: "WARNING: Permissions may only be disabled, once disabled they may not be reactivated!!",
-            issued_assets: "Issued Assets"
+            issued_assets: "Issued Assets",
+            short: "Short name (max 32 characters)",
+            condition: "Condition (max 60 characters)",
+            expiry: "Market resolution date",
+            decimals: "Number of decimal points",
+            mpa: "SmartCoin",
+            pm: "Binary prediction market",
+            bitasset_opts: "SmartCoin options",
+            feed_lifetime_sec: "Feed lifetime in minutes",
+            minimum_feeds: "Minimum number of feeds",
+            force_settlement_delay_sec: "Delay for forced settlements (minutes)",
+            force_settlement_offset_percent: "Percent offset of forced settlements",
+            maximum_force_settlement_volume: "Max force settle volume (percent)",
+            backing: "Short backing asset",
+            error_precision: "That asset does not have the same precision as %(asset)s",
+            error_invalid: "That asset may not be used",
+            market: "Preferred market pairing"
         },
         connections: {
             known: "Known by",
@@ -456,7 +474,7 @@
             membership_annual_fee : "Annual Membership",
             membership_lifetime_fee : "Lifetime Membership",
             symbol3 : "Symbols with 3 Characters",
-            symbol4 : "Sybmols with 4 Characters",
+            symbol4 : "Symbols with 4 Characters",
             long_symbol : "Longer Symbols"
         },
         whitelist_states: {
@@ -630,20 +648,28 @@
         yes: "Yes",
         no: "No",
         walletLockTimeout: "Wallet auto-lock time (seconds)",
-        multiAccountMode: "Multiple Accounts Mode"
+        multiAccountMode: "Multiple Accounts Mode",
+        themes: "Theme",
+        "darkTheme": "Dark theme",
+        "lightTheme": "Light theme",
+        "olDarkTheme": "Openledger dark"
     },
     footer: {
         title: "Graphene",
         block: "Head block",
-        loading: "Loading..."
+        loading: "Loading...",
+        backup: "Backup Required",
+        nosync: "Blockchain is out of sync, please wait until it's synchronized..",
+        connection: "No Blockchain connection",
+        brainkey: "Backup brainkey recommended"
     },
     exchange: {
         market: "Market",
         price_history: "Price Chart",
         order_depth: "Market Depth",
-        history: "All History",
-        my_history: "My History",
-        balance: "Your balance",
+        history: "Market trades",
+        my_history: "My trades",
+        balance: "Balance",
         lowest_ask: "Lowest ask",
         highest_bid: "Highest bid",
         total: "Total",
@@ -670,7 +696,7 @@
         market_name: "My Markets",
         quote_supply: "Quote supply",
         base_supply: "Base supply",
-        more: "All markets",
+        more: "Find markets",
         volume_24: "24hr Volume",
         change: "Change",
         confirm: "Your order is %(diff)s% away from the current price, are you sure?",
@@ -685,15 +711,26 @@
         index: "Index",
         my_bids: "My bids",
         my_asks: "My asks",
+        my_orders: "My open orders",
         settle_orders: "Settle orders",
-        asks: "Asks",
-        bids: "Bids",
+        asks: "Sell orders",
+        bids: "Buy orders",
         no_data: "No data",
-        time: "Time interval",
+        time: "Candlesticks",
+        zoom: "Zoom",
         borrow: "Borrow",
         no_balance: "Insufficient balance",
         invalid_amount: "Invalid amount",
-        invalid_price: "Invalid price"
+        invalid_price: "Invalid price",
+        show_bids: "Show all bids",
+        show_asks: "Show all asks",
+        hide: "Hide",
+        short: "Short",
+        others: "Others",
+        invert: "Invert the price",
+        to_market: "Go to market",
+        zoom_all: "All",
+        settings: "Chart options"
     },
     fees: {
      title: "Fee Schedule"
@@ -864,6 +901,12 @@
         title: "Request settlement of %(asset)s",
         amount: "Amount to settle",
         submit: "Settle asset"
+      },
+      reserve: {
+        title: "Reserve (burn) asset",
+        amount: "Amount to reserve",
+        from: "Account to reserve from",
+        submit: "Reserve"
       },
       ok: "OK"
     },
