@@ -325,7 +325,9 @@ class AccountVoting extends React.Component {
                         </button>) : null}
                 </div>
 
-                <Tabs setting="votingTab" style={{maxWidth: "800px"}}>
+
+                <div className="exchange-bordered">
+                <Tabs setting="votingTab" tabsClass="no-padding bordered-header" contentClass="grid-content">
 
                         <Tab title="account.votes.proxy_short">
                             <div className="content-block">
@@ -405,6 +407,7 @@ class AccountVoting extends React.Component {
                             </div>
                         </Tab>
                 </Tabs>
+                </div>
             </div>
         )
     }
