@@ -44,6 +44,8 @@ class CreateAccount extends React.Component {
             show_identicon: false
         };
         this.onFinishConfirm = this.onFinishConfirm.bind(this);
+
+        this.accountNameInput = null;
     }
 
     shouldComponentUpdate(nextProps, nextState) {
@@ -227,7 +229,7 @@ class CreateAccount extends React.Component {
                     <div className="content-block center-content">
                         <div style={{width: '21em'}}>
                             <form onSubmit={this.onSubmit.bind(this)} noValidate>
-                                <AccountNameInput ref="account_name" cheapNameOnly={first_account}
+                                <AccountNameInput ref={(ref) => {if (ref) {this.accountNameInput = ref.refs.nameInput;}}} cheapNameOnly={first_account}
                                                   onChange={this.onAccountNameChange.bind(this)}
                                                   accountShouldNotExist={true} focus={true}/>
 
