@@ -29,13 +29,13 @@ class TranswiserDepositWithdraw extends React.Component {
     }
 
     requestDepositUrl(){
+        let pair = this.depositCoin.toLocaleLowerCase() + ":" + this.props.receiveAsset.get('symbol').toLocaleLowerCase();
 
         fetch( this.apiUrl, {
             method:'get',
             headers: new Headers( { "Accept": "application/json", "Content-Type":"application/json" } )
         }).then( reply => { reply.json().then( json => {
                 // console.log( "reply: ", json )
-                let pair = this.depositCoin.toLocaleLowerCase() + ":" + this.props.receiveAsset.get('symbol').toLocaleLowerCase();
                 let setting = json[pair];
                 if( setting )
                     this.setState({
@@ -66,6 +66,7 @@ class TranswiserDepositWithdraw extends React.Component {
     }
 
     onWithdraw() {
+        // console.log('onWithdraw', this.getWithdrawModalId());
         ZfApi.publish(this.getWithdrawModalId(), "open");
     }
 
@@ -74,9 +75,13 @@ class TranswiserDepositWithdraw extends React.Component {
     }
 
     render() {
+        let loading = (<tr style={{display:"block"}}><td>loading</td><td></td><td></td><td></td></tr>);
+
+        if( !this.props.account || !this.props.issuerAccount || !this.props.receiveAsset )
+            return loading;
+
         if (!this.state.depositUrl) {
-            this.requestDepositUrl()
-            return <tr><td></td><td></td><td></td><td></td></tr>
+            this.requestDepositUrl(); return loading;
         }
 
         let withdrawModalId = this.getWithdrawModalId();
@@ -86,8 +91,6 @@ class TranswiserDepositWithdraw extends React.Component {
             <td>{this.props.receiveAsset.get('symbol')} </td>
             <td>
                 <button className={"button outline"} onClick={this.onDeposit.bind(this)}> <Translate content="gateway.deposit" /> </button>
-        {
-            //*
                 <Modal id={depositModalId} overlay={true}>
                     <Trigger close={depositModalId}>
                         <a href="#" className="close-button">&times;</a>
@@ -103,8 +106,6 @@ class TranswiserDepositWithdraw extends React.Component {
                             modalId={depositModalId} />
                     </div>
                 </Modal>
-            //*/
-                        }
             </td>
             <td> <AccountBalance account={this.props.account.get('name')} asset={this.props.receiveAsset.get('symbol')} /> </td>
             <td>
