@@ -10,6 +10,8 @@ import FormattedAsset from "../Utility/FormattedAsset";
 import EquivalentValueComponent from "../Utility/EquivalentValueComponent";
 import {chain_types} from "@graphene/chain";
 let {operations} = chain_types;
+import ChainStore from "api/ChainStore";
+
 let ops = Object.keys(operations);
 
 // Define groups and their corresponding operation ids
@@ -19,7 +21,10 @@ let fee_grouping = {
     market   : [1,2,3,4,17,18],
     account  : [5,6,7,8,9],
     business : [20,21,22,23,24,29,30,31,34,35,36],
-}
+};
+
+// Operations that require LTM
+let ltm_required = [5, 7, 20, 21, 34];
 
 @BindToChainState({keep_updating:true})
 class FeeGroup extends React.Component {
@@ -45,11 +50,12 @@ class FeeGroup extends React.Component {
     render() {
         let {globalObject, settings, opIds} = this.props;
         globalObject = globalObject.toJSON();
+        const core_asset = ChainStore.getAsset("1.3.0");
 
         let current_fees = globalObject.parameters.current_fees;
         let scale   = current_fees.scale;
         let feesRaw = current_fees.parameters;
-        let preferredUnit = settings.get("unit") || "BTS";
+        let preferredUnit = settings.get("unit") || core_asset.get("symbol");
 
         let trxTypes = counterpart.translate("transaction.trxTypes");
 
@@ -88,18 +94,27 @@ class FeeGroup extends React.Component {
                              </td>)
                 }
 
-                rows.push(
-                        <tr key={opId + key}>
-                            {title}
-                            <td>{feeTypes[key]}</td>
-                            <td style={{textAlign: "right"}}>{equivalentAmount}</td>
-                            <td style={{textAlign: "right"}}>{equivalentAmountLTM}</td>
-                        </tr>
-                       );
+                if (ltm_required.indexOf(opId)<0) {
+                    rows.push(
+                            <tr key={opId.toString() + key}>
+                                {title}
+                                <td>{feeTypes[key]}</td>
+                                <td style={{textAlign: "right"}}>{equivalentAmount}</td>
+                                <td style={{textAlign: "right"}}>{equivalentAmountLTM}</td>
+                            </tr>
+                           );
+                } else {
+                    rows.push(
+                            <tr key={opId.toString() + key}>
+                                {title}
+                                <td>{feeTypes[key]}</td>
+                                <td style={{textAlign: "right"}}>- <sup>*</sup></td>
+                                <td style={{textAlign: "right"}}>{equivalentAmountLTM}</td>
+                            </tr>
+                           );
+                }
             }
-
             return (<tbody>{rows}</tbody>);
-
         })
 
         return (   
@@ -107,7 +122,7 @@ class FeeGroup extends React.Component {
                     <div className="card-divider">{this.props.title}</div>
                     <table className="table">
                      <thead>
-                      <tr>
+                      <tr key={this.props.title}>
                        <th><Translate content={"explorer.block.op"} /></th>
                        <th><Translate content={"explorer.fees.type"} /></th>
                        <th style={{textAlign: "right"}}><Translate content={"explorer.fees.fee"} /></th>
@@ -119,14 +134,12 @@ class FeeGroup extends React.Component {
                    </div>
            );
     }
-
 }
 
 class Fees extends React.Component {
 
     render() {
 
-        let arr = [5, 6, 7, 8, 9];
         let FeeGroupsTitle  = counterpart.translate("transaction.feeGroups");
         let feeGroups = []
 
