@@ -8,6 +8,7 @@ import MarketHistory from "./MarketHistory";
 import MyMarkets from "./MyMarkets";
 import BuySell from "./BuySell";
 import utils from "common/utils";
+import assetUtils from "common/asset_utils";
 import PriceChart from "./PriceChart";
 import DepthHighChart from "./DepthHighChart";
 import {debounce, cloneDeep} from "lodash";
@@ -31,11 +32,13 @@ import LoadingIndicator from "../LoadingIndicator";
 import ConfirmOrderModal from "./ConfirmOrderModal";
 import IndicatorModal from "./IndicatorModal";
 import OpenSettleOrders from "./OpenSettleOrders";
+import counterpart from "counterpart";
 
 require("./exchange.scss");
 
 let emitter = ee.emitter();
-let callListener, limitListener, newCallListener;
+let callListener, limitListener, newCallListener, feedUpdateListener, settleOrderListener;
+let SATOSHI = 8;
 
 Highcharts.setOptions({
     global: {
@@ -99,11 +102,17 @@ class Exchange extends React.Component {
         super();
 
         this.state = this._initialState(props);
+
+        this._getWindowSize = debounce(this._getWindowSize.bind(this), 150);
     }
 
+
     _initialState(props) {
+        let ws = props.viewSettings;
+
         return {
             history: [],
+            buySellOpen: ws.get("buySellOpen", true),
             buyAmount: 0,
             displaySellPrice: 0,
             displayBuyPrice: 0,
@@ -131,19 +140,21 @@ class Exchange extends React.Component {
             sellAmount: 0,
             sellTotal: 0,
             sub: null,
-            flipBuySell: props.viewSettings.get("flipBuySell"),
+            flipBuySell: ws.get("flipBuySell", false),
             favorite: false,
-            showDepthChart: props.viewSettings.get("showDepthChart"),
-            leftOrderBook: props.viewSettings.get("leftOrderBook"),
+            showDepthChart: ws.get("showDepthChart", false),
+            leftOrderBook: ws.get("leftOrderBook", false),
             buyDiff: false,
             sellDiff: false,
-            indicators: props.viewSettings.get("indicators") || {
+            indicators: ws.get("indicators", {
                 rsi: false,
-                sma: false,
+                sma: true,
                 atr: false,
-                ema: false
-            },
-            indicatorSettings: props.viewSettings.get("indicatorSettings") || {
+                ema: true
+            }),
+            preferCoreBuyFee: ws.get("preferCoreBuyFee", true) ,
+            preferCoreSellFee: ws.get("preferCoreSellFee", true),
+            indicatorSettings: ws.get("indicatorSettings") || {
                 rsi: {
                     period: 14,
                     overbought: 70,
@@ -159,7 +170,11 @@ class Exchange extends React.Component {
                     period: 10,
                     index: 0
                 }
-            }
+            },
+            height: window.innerHeight,
+            width: window.innerWidth,
+            chartHeight: ws.get("chartHeight", 425),
+            currentPeriod: ws.get("currentPeriod", 3600 * 24 * 7)
         };
     }
 
@@ -205,14 +220,39 @@ class Exchange extends React.Component {
         emitter.on('cancel-order', limitListener = MarketsActions.cancelLimitOrderSuccess);
         emitter.on('close-call', callListener = MarketsActions.closeCallOrderSuccess);
         emitter.on('call-order-update', newCallListener = MarketsActions.callOrderUpdate);
+        emitter.on('bitasset-update', feedUpdateListener = MarketsActions.feedUpdate);
+        emitter.on('settle-order-update', settleOrderListener = (object) => {
+            let {isMarketAsset, marketAsset} = market_utils.isMarketAsset(this.props.quoteAsset, this.props.baseAsset);
+            console.log("settle-order-update:", object, "isMarketAsset:", isMarketAsset, "marketAsset:", marketAsset);
+
+            if (isMarketAsset && marketAsset.id === object.balance.asset_id) {
+               MarketsActions.settleOrderUpdate(marketAsset.id);
+            }
+        });
+
+        window.addEventListener("resize", this._getWindowSize, false);        
     }
 
     componentDidMount() {
         let centerContainer = ReactDOM.findDOMNode(this.refs.center);
-        Ps.initialize(centerContainer);
+        if (centerContainer) {
+            Ps.initialize(centerContainer);
+        }
         SettingsActions.changeViewSetting({
             lastMarket: this.props.quoteAsset.get("symbol") + "_" + this.props.baseAsset.get("symbol")
         });
+
+    }
+
+
+    _getWindowSize() {
+        let {innerHeight, innerWidth} = window;
+        if (innerHeight !== this.state.height || innerWidth !== this.state.width) {
+            this.setState({
+                height: innerHeight,
+                width: innerWidth
+            });
+        }
     }
 
     _addMarket(quote, base) {
@@ -225,6 +265,12 @@ class Exchange extends React.Component {
     }
 
     componentWillReceiveProps(nextProps) {
+        if (nextProps.baseAsset && nextProps.baseAsset.getIn(["bitasset", "is_prediction_market"])) {
+            console.log(nextProps.baseAsset.get("symbol"), "is prediction market");
+            // console.log("this.props:", this.props);
+            this.props.history.push(`market/${nextProps.baseAsset.get("symbol")}_${nextProps.quoteAsset.get("symbol")}`)
+        }
+
         if (nextProps.quoteAsset.toJS && nextProps.baseAsset.toJS) {
             // this._addMarket(nextProps.quoteAsset.get("symbol"), nextProps.baseAsset.get("symbol"));
             if (!this.state.sub) {
@@ -250,18 +296,24 @@ class Exchange extends React.Component {
         emitter.off('cancel-order', limitListener);
         emitter.off('close-call', callListener);
         emitter.off('call-order-update', newCallListener);
+        emitter.off('bitasset-update', feedUpdateListener);
+
+        window.removeEventListener("resize", this._getWindowSize, false);
+
     }
 
-    _createLimitOrder(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID) {
-        console.log("createLimitOrder:", buyAssetAmount, sellAssetAmount);
+    _createPredictionShort(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID) {
+        console.log("createPredictionShort:", buyAssetAmount, sellAssetAmount);
+        let coreAsset = ChainStore.getAsset("1.3.0");
         let expiration = new Date();
         // TODO: Add selector for expiry
         expiration.setYear(expiration.getFullYear() + 5);
-        MarketsActions.createLimitOrder(
+        MarketsActions.createPredictionShort(
             this.props.currentAccount.get("id"),
             utils.get_satoshi_amount(sellAssetAmount, sellAsset),
             sellAsset,
             utils.get_satoshi_amount(buyAssetAmount, buyAsset),
+            utils.get_satoshi_amount(sellAssetAmount, coreAsset),
             buyAsset,
             expiration,
             false, // fill or kill TODO: add fill or kill switch
@@ -280,7 +332,7 @@ class Exchange extends React.Component {
     _getFee(asset) {
         let fee = utils.estimateFee("limit_order_create", [], ChainStore.getObject("2.0.0")) || 0;
 
-        if (!asset) return fee;
+        if (!asset || asset.get("id") === "1.3.0") return fee;
         let cer = asset.getIn(["options", "core_exchange_rate"]).toJS();
         let coreAsset = ChainStore.getAsset("1.3.0");
         if (!coreAsset) return 0;
@@ -291,23 +343,30 @@ class Exchange extends React.Component {
         return eqValue;
     }
 
-    _verifyFee(fee, feeAsset, sellAmount, sellBalance, coreBalance) {
+    _verifyFee(fee, feeAsset, sellAsset, sellAmount, sellBalance, coreBalance) {
         let coreFee = this._getFee();
 
-        let sellPrecision = utils.get_asset_precision(feeAsset);
+        let sellPrecision = utils.get_asset_precision(sellAsset);
         let sellSum = fee + parseInt(sellAmount * sellPrecision, 10);
-        if (sellSum <= sellBalance) { // Sufficient balance in asset to pay fee
-            return feeAsset.get("id");
-        } else if (coreFee <= coreBalance && feeAsset.get("id") !== "1.3.0") { // Sufficient balance in core asset to pay fee
-            return "1.3.0";
+        if (feeAsset.get("id") === "1.3.0") {
+            if (coreFee <= coreBalance) {
+                return "1.3.0";
+            } else {
+                return null;
+            }
         } else {
-            return null; // Unable to pay fee in either asset
+            if (sellSum <= sellBalance) { // Sufficient balance in asset to pay fee
+                return feeAsset.get("id");
+            } else if (coreFee <= coreBalance && feeAsset.get("id") !== "1.3.0") { // Sufficient balance in core asset to pay fee
+                return "1.3.0";
+            } else {
+                return null; // Unable to pay fee in either asset
+            }
         }
     }
 
-    _createLimitOrderConfirm(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, sellBalance, coreBalance, type, e) {
+    _createLimitOrderConfirm(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, sellBalance, coreBalance, feeAsset, type, short = true, e) {
         e.preventDefault();
-        let feeID; // = "1.3.0";
 
         sellBalance = sellBalance ? parseInt(ChainStore.getObject(sellBalance).toJS().balance, 10) : 0;
         coreBalance = coreBalance ? parseInt(ChainStore.getObject(coreBalance).toJS().balance, 10) : 0;
@@ -316,13 +375,13 @@ class Exchange extends React.Component {
 
         // Convert fee to relevant asset fee and check if user has sufficient balance
         let feeAmount;
-        if (sellAsset.get("id") !== "1.3.0") {
-            feeAmount = this._getFee(sellAsset);
-        } else {
-            feeAmount = this._getFee();
-        }
+        // if (sellAsset.get("id") !== "1.3.0") {
+        //     feeAmount = this._getFee(sellAsset);
+        // } else {
+        feeAmount = this._getFee(feeAsset);
+        // }
             
-        feeID = this._verifyFee(feeAmount, sellAsset, sellAssetAmount, sellBalance, coreBalance);
+        let feeID = this._verifyFee(feeAmount, feeAsset, sellAsset, sellAssetAmount, sellBalance, coreBalance);
 
         if (!feeID) {
             return notify.addNotification({
@@ -349,7 +408,9 @@ class Exchange extends React.Component {
             }
         }
 
-        if ((sellAssetAmount * sellPrecision) > sellBalance) {
+        let isPredictionMarket = sellAsset.getIn(["bitasset", "is_prediction_market"]);
+
+        if ((sellAssetAmount * sellPrecision) > sellBalance && !isPredictionMarket) {
             return notify.addNotification({
                 message: "Insufficient funds to place order. Required: " + sellAssetAmount + " " + sellAsset.get("symbol"),
                 level: "error"
@@ -363,7 +424,35 @@ class Exchange extends React.Component {
             });
         }
 
+        if (type === "sell" && isPredictionMarket && short) {
+            return this._createPredictionShort(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID);
+        }
         this._createLimitOrder(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID);
+    }
+
+    _createLimitOrder(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, feeID) {
+        console.log("createLimitOrder:", buyAssetAmount, sellAssetAmount);
+        let expiration = new Date();
+        // TODO: Add selector for expiry
+        expiration.setYear(expiration.getFullYear() + 5);
+        MarketsActions.createLimitOrder(
+            this.props.currentAccount.get("id"),
+            utils.get_satoshi_amount(sellAssetAmount, sellAsset),
+            sellAsset,
+            utils.get_satoshi_amount(buyAssetAmount, buyAsset),
+            buyAsset,
+            expiration,
+            false, // fill or kill TODO: add fill or kill switch
+            feeID
+        ).then(result => {
+            if (result.error) {
+                if (result.error.message !== "wallet locked")
+                    notify.addNotification({
+                        message: "Unknown error. Failed to place order for " + buyAssetAmount + " " + buyAsset.symbol,
+                        level: "error"
+                    });
+            }
+        });
     }
 
     _forceBuy(buyAsset, sellAsset, buyAssetAmount, sellAssetAmount, value, sellBalance, coreBalance) {
@@ -416,37 +505,46 @@ class Exchange extends React.Component {
         }
     }
 
+    _changeZoomPeriod(size, e) {
+        e.preventDefault();
+        if (size !== this.state.currentPeriod) {
+            this.setState({
+                currentPeriod: size
+            });
+            SettingsActions.changeViewSetting({
+                currentPeriod: size
+            });
+        }
+    }    
+
     _subToMarket(props, newBucketSize) {
         let {quoteAsset, baseAsset, bucketSize} = props;
         if (newBucketSize) {
             bucketSize = newBucketSize;
         }
         if (quoteAsset.get("id") && baseAsset.get("id")) {
-            MarketsActions.subscribeMarket(baseAsset, quoteAsset, bucketSize);
+            MarketsActions.subscribeMarket.defer(baseAsset, quoteAsset, bucketSize);
             this.setState({sub: `${quoteAsset.get("id")}_${baseAsset.get("id")}`});
         }
     }
 
     _depthChartClick(base, quote, power, e) {
         e.preventDefault();
-        let value = this._limitByPrecision(e.xAxis[0].value / power, quote);
+        let value = market_utils.limitByPrecision(e.xAxis[0].value / power, quote);
         let buyPrice = this._getBuyPrice(e.xAxis[0].value / power);
         let sellPrice = this._getSellPrice(e.xAxis[0].value / power);
         let displayBuyPrice = this._getDisplayPrice("bid", buyPrice);
         let displaySellPrice = this._getDisplayPrice("ask", sellPrice);
-        // let buyPrice = this._buyPriceChanged(base, quote, {target: {value: value}});
 
         this.setState({
             depthLine: value,
             buyPrice: buyPrice,
             displayBuyPrice: displayBuyPrice,
-            buyTotal: this._limitByPrecision(this.getBuyTotal(buyPrice, this.state.buyAmount), base),
+            buyTotal: market_utils.limitByPrecision(this.getBuyTotal(buyPrice, this.state.buyAmount), base),
             sellPrice: sellPrice,
             displaySellPrice: displaySellPrice,
-            sellTotal: this._limitByPrecision(this.getSellTotal(sellPrice, this.state.sellAmount), base)
+            sellTotal: market_utils.limitByPrecision(this.getSellTotal(sellPrice, this.state.sellAmount), base)
         });
-
-        // this._sellPriceChanged(base, quote, {target: {value: value}});
     }
 
     _addZero(value) {
@@ -460,29 +558,38 @@ class Exchange extends React.Component {
         return value;
     }
 
-    _limitByPrecision(value, asset, floor = true) {
-        let assetPrecision = asset.toJS ? asset.get("precision") : asset.precision;
-        let valueString = value.toString();
-        let splitString = valueString.split(".");
-        if (splitString.length === 1 || splitString.length === 2 && splitString[1].length <= assetPrecision) {
-            return value;
-        }
-        let precision = utils.get_asset_precision(assetPrecision);
-        value = floor ? Math.floor(value * precision) / precision : Math.round(value * precision) / precision;
-        if (isNaN(value) || !isFinite(value)) {
-            return 0;
-        }
-        return value;
-    }
+    _
 
     _buyPriceChanged(base, quote, e) {
-        let amount = this._limitByPrecision(e.target.value, base);
+
+        let split = e.target.value.split(".");
+        if (split.length === 2 && split[1].length === SATOSHI + 1) {
+            return;
+        }
+
+        let amount = market_utils.limitByPrecision(e.target.value, {precision: SATOSHI});
         let price = this._getBuyPrice(amount);
 
         this.setState({
             buyPrice: price,
             displayBuyPrice: amount,
-            buyTotal: this._limitByPrecision(this.getBuyTotal(price, this.state.buyAmount), base),
+            buyTotal: market_utils.limitByPrecision(this.getBuyTotal(price, this.state.buyAmount), base),
+            depthLine: amount
+        });
+    }
+
+    _sellPriceChanged(base, quote, e) {
+        let split = e.target.value.split(".");
+        if (split.length === 2 && split[1].length === SATOSHI + 1) {
+            return;
+        }
+        let amount = market_utils.limitByPrecision(e.target.value, {precision: SATOSHI});
+        let price = this._getSellPrice(amount);
+
+        this.setState({
+            sellPrice: price,
+            displaySellPrice: amount,
+            sellTotal: market_utils.limitByPrecision(this.getSellTotal(price, this.state.sellAmount), base),
             depthLine: amount
         });
     }
@@ -490,25 +597,26 @@ class Exchange extends React.Component {
     _buyAmountChanged(base, quote, e) {
         let value = e.target.value;
         if (e.target.value.indexOf(".") !== e.target.value.length -1) {
-            value = this._limitByPrecision(e.target.value, quote);
+            value = market_utils.limitByPrecision(e.target.value, quote);
         }
 
         this.setState({
             buyAmount: this._addZero(value),
-            buyTotal: this._limitByPrecision(this.getBuyTotal(this.state.buyPrice, value), base)
+            buyTotal: market_utils.limitByPrecision(this.getBuyTotal(this.state.buyPrice, value), base)
         });
     }
 
     _buyTotalChanged(base, quote, e) {
         let value = e.target.value;
         if (e.target.value.indexOf(".") !== e.target.value.length -1) {
-            value = this._limitByPrecision(e.target.value, base);
+            value = market_utils.limitByPrecision(e.target.value, base);
         }
+
 
         let amount = this.getBuyAmount(this.state.buyPrice, value);
 
         this.setState({
-            buyAmount: this._limitByPrecision(amount, quote),
+            buyAmount: market_utils.limitByPrecision(amount, quote),
             buyTotal: this._addZero(value)
         });
     }
@@ -516,34 +624,22 @@ class Exchange extends React.Component {
     _sellAmountChanged(base, quote, e) {
         let value = e.target.value;
         if (e.target.value.indexOf(".") !== e.target.value.length -1) {
-            value = this._limitByPrecision(e.target.value, quote);
+            value = market_utils.limitByPrecision(e.target.value, quote);
         }
         this.setState({
             sellAmount: this._addZero(value),
-            sellTotal: this._limitByPrecision(this.getSellTotal(this.state.sellPrice, value), base)
-        });
-    }
-
-    _sellPriceChanged(base, quote, e) {
-        let amount = this._limitByPrecision(e.target.value, base);
-        let price = this._getSellPrice(amount);
-
-        this.setState({
-            sellPrice: price,
-            displaySellPrice: amount,
-            sellTotal: this._limitByPrecision(this.getSellTotal(price, this.state.sellAmount), base),
-            depthLine: amount
+            sellTotal: market_utils.limitByPrecision(this.getSellTotal(this.state.sellPrice, value), base)
         });
     }
 
     _sellTotalChanged(base, quote, e) {
         let value = e.target.value;
         if (e.target.value.indexOf(".") !== e.target.value.length -1) {
-            value = this._limitByPrecision(e.target.value, base);
+            value = market_utils.limitByPrecision(e.target.value, base);
         }
 
         this.setState({
-            sellAmount: this._limitByPrecision(this.getSellAmount(this.state.sellPrice, value), quote),
+            sellAmount: market_utils.limitByPrecision(this.getSellAmount(this.state.sellPrice, value), quote),
             sellTotal: this._addZero(value)
         });
     }
@@ -556,9 +652,19 @@ class Exchange extends React.Component {
         this.setState({flipBuySell: !this.state.flipBuySell});
     }
 
-    getSellAmount(price, total = 0) {
+    _toggleOpenBuySell() {
+        SettingsActions.changeViewSetting({
+            buySellOpen: !this.state.buySellOpen
+        });
+
+        this.setState({buySellOpen: !this.state.buySellOpen});
+    }
+
+    getSellAmount(price, total = 0, satAmount) {
         let amountPrecision = utils.get_asset_precision(this.props.quoteAsset.get("precision"));
-        let satAmount = utils.get_satoshi_amount(total, this.props.baseAsset);    
+        if (!satAmount) {
+            satAmount = utils.get_satoshi_amount(total, this.props.baseAsset);    
+        } 
         return ((satAmount / price.base.amount) * price.quote.amount) / amountPrecision;
     }
 
@@ -568,17 +674,22 @@ class Exchange extends React.Component {
         return ((satAmount / price.quote.amount) * price.base.amount) / totalPrecision;
     }
 
-    getBuyAmount(price, total = 0) {
+    getBuyAmount(price, total = 0, satAmount) {
         let amountPrecision = utils.get_asset_precision(this.props.quoteAsset.get("precision"));
-        let satAmount = utils.get_satoshi_amount(total, this.props.baseAsset);
+        if (!satAmount) {
+            satAmount = utils.get_satoshi_amount(total, this.props.baseAsset);
+        }
 
         return ((satAmount / price.quote.amount) * price.base.amount) / amountPrecision;
     }
 
-    getBuyTotal(price, amount = 0) {
+    getBuyTotal(price, amount = 0, satAmount) {
         let totalPrecision = utils.get_asset_precision(this.props.baseAsset.get("precision"));
-        let satAmount = utils.get_satoshi_amount(amount, this.props.quoteAsset);
-        return ((satAmount / price.base.amount) * price.quote.amount) / totalPrecision;
+        if (!satAmount) {
+            satAmount = utils.get_satoshi_amount(amount, this.props.quoteAsset);    
+        }
+
+        return (Math.floor(0.5 + (satAmount / price.base.amount) * price.quote.amount)) / totalPrecision;
     }
 
     _toggleCharts() {
@@ -603,9 +714,9 @@ class Exchange extends React.Component {
             let {buyTotal, buyAmount} = this.state;
 
             if (buyAmount) {
-                buyTotal = this._limitByPrecision(this.getBuyTotal(price, this.state.buyAmount), base);
+                buyTotal = market_utils.limitByPrecision(this.getBuyTotal(price, this.state.buyAmount), base);
             } else if (buyTotal) {
-                buyAmount = this._limitByPrecision(this.getBuyAmount(price, buyTotal), quote);
+                buyAmount = market_utils.limitByPrecision(this.getBuyAmount(price, buyTotal), quote);
             }
 
             this.setState({
@@ -620,9 +731,9 @@ class Exchange extends React.Component {
             let {sellTotal, sellAmount} = this.state;
 
             if (sellAmount) {
-                sellTotal = this._limitByPrecision(this.getSellTotal(price, this.state.sellAmount), base);
+                sellTotal = market_utils.limitByPrecision(this.getSellTotal(price, this.state.sellAmount), base);
             } else if (sellTotal) {
-                sellAmount = this._limitByPrecision(this.getSellAmount(price, sellTotal), quote);
+                sellAmount = market_utils.limitByPrecision(this.getSellAmount(price, sellTotal), quote);
             }
             this.setState({
                 sellPrice: price,
@@ -637,39 +748,50 @@ class Exchange extends React.Component {
 
     _orderbookClick(base, quote, type, order) {
         let precision = utils.get_asset_precision(quote.get("precision") + base.get("precision"));
-        if (type === "bid") {
-            let value = order.totalAmount.toString();
-            if (value.indexOf(".") !== value.length -1) {
-                value = this._limitByPrecision(order.totalAmount, quote);
-            }
+        
+        let value = order.amount.toString();
+        if (value.indexOf(".") !== value.length - 1) {
+            value = market_utils.limitByPrecision(order.amount, quote);
+        }
 
+        if (type === "bid") {
 
             let displaySellPrice = this._getDisplayPrice("ask", order.sell_price);
-
-            let total = this.getSellTotal(order.sell_price, value);
+            let sellAmount = market_utils.limitByPrecision(this.getSellAmount(order.sell_price, null, order.totalForSale), quote);
 
             this.setState({
                 displaySellPrice: displaySellPrice,
                 sellPrice: order.sell_price,
-                sellAmount: value,
-                sellTotal: this._limitByPrecision(total, base)
+                sellAmount: sellAmount,
+                sellTotal: utils.get_asset_amount(order.totalForSale, base),
+                displayBuyPrice: displaySellPrice,
+                buyPrice: {
+                    quote: order.sell_price.base,
+                    base: order.sell_price.quote
+                },
+                buyAmount: null,
+                buyTotal: null
             });
 
         } else if (type === "ask") {
-            let value = order.totalAmount.toString();
-            if (value.indexOf(".") !== value.length -1) {
-                value = this._limitByPrecision(order.totalAmount, base);
-            }
 
             let displayBuyPrice = this._getDisplayPrice("bid", order.sell_price);
-
-            let total = this.getBuyTotal(order.sell_price, value);
+            // Calculate total
+            // let total = this.getBuyTotal(order.sell_price, null, order.totalForSale);
+            let buyAmount = market_utils.limitByPrecision(this.getBuyAmount(order.sell_price, null, order.totalValue), quote);
 
             this.setState({
                 displayBuyPrice: displayBuyPrice,
                 buyPrice: order.sell_price,
-                buyAmount: value,
-                buyTotal: this._limitByPrecision(total, base)
+                buyAmount: buyAmount,
+                buyTotal: market_utils.limitByPrecision(utils.get_asset_amount(order.totalValue, base), base),
+                displaySellPrice: displayBuyPrice,
+                sellPrice: {
+                    quote: order.sell_price.base,
+                    base: order.sell_price.quote
+                },
+                sellAmount: null,
+                sellTotal: null
             });
         }
     }
@@ -689,37 +811,37 @@ class Exchange extends React.Component {
     _getBuyPrice(price) {
         let nominator = utils.get_satoshi_amount(price, this.props.baseAsset)
         let denominator = utils.get_satoshi_amount(1, this.props.quoteAsset);
-        
+
+        let integerRatio = market_utils.priceToObject(price, "ask");
         let {baseAsset, quoteAsset} = this.props;
         let quotePrecision = utils.get_asset_precision(quoteAsset.get("precision"));
         let basePrecision = utils.get_asset_precision(baseAsset.get("precision"));
 
         return {
-            base: {
-                 asset_id: baseAsset.get("id"),
-                 amount: denominator
-            },
             quote: {
+                 asset_id: baseAsset.get("id"),
+                 amount: integerRatio.base * basePrecision
+            },
+            base: {
                 asset_id: quoteAsset.get("id"),
-                amount: nominator
+                amount: integerRatio.quote * quotePrecision
             }
         };
     }
 
     _getDisplayPrice(type, priceObject) {
         let {quoteAsset, baseAsset} = this.props;
-        let precision =  Math.min(8, quoteAsset.get("precision") + baseAsset.get("precision"));
         let price;
 
         switch (type) {
             case "bid":
                 price = utils.get_asset_price(priceObject.quote.amount, baseAsset, priceObject.base.amount, quoteAsset);
-                price = this._limitByPrecision(this._addZero(price), {precision}, false);
+                price = market_utils.limitByPrecision(this._addZero(price), {precision: SATOSHI}, false);
                 return isNaN(price) ? 0 : price;
 
             case "ask":
                 price = utils.get_asset_price(priceObject.base.amount, baseAsset, priceObject.quote.amount, quoteAsset);
-                price = this._limitByPrecision(this._addZero(price), {precision}, false);
+                price = market_utils.limitByPrecision(this._addZero(price), {precision: SATOSHI}, false);
                 return isNaN(price) ? 0 : price;
 
             default:
@@ -733,26 +855,28 @@ class Exchange extends React.Component {
         let nominator = utils.get_satoshi_amount(price, this.props.baseAsset)
         let denominator = utils.get_satoshi_amount(1, this.props.quoteAsset);
         
+        let integerRatio = market_utils.priceToObject(price, "bid");
         let {baseAsset, quoteAsset} = this.props;
-
+        
         let quotePrecision = utils.get_asset_precision(quoteAsset.get("precision"));
         let basePrecision = utils.get_asset_precision(baseAsset.get("precision"));
 
         return {
-            base: {
-                 asset_id: this.props.quoteAsset.get("id"),
-                 amount: nominator
-            },
             quote: {
-                asset_id: this.props.baseAsset.get("id"),
-                amount: denominator
+                 asset_id: baseAsset.get("id"),
+                 amount: integerRatio.base * quotePrecision
+            },
+            base: {
+                asset_id: quoteAsset.get("id"),
+                amount: integerRatio.quote * basePrecision
             }
         };
     }
 
     _parseMarket() {
         let {bids, asks, calls, invertedCalls} = this.props;
-        let {showCallLimit} = this._getSettlementInfo;
+
+        let {showCallLimit} = this._getSettlementInfo();
         let combinedAsks, combinedBids, highestBid, lowestAsk;
 
         if (calls.length && invertedCalls) {
@@ -794,6 +918,7 @@ class Exchange extends React.Component {
 
     _getSettlementInfo() {
         let {quoteAsset: quote, baseAsset: base, bids, asks, lowestCallPrice} = this.props;
+
         let settlement_price, core_rate, short_squeeze, flipped,
             settlementBase, settlementQuote, settlementPrice, highestBid,
             squeezePrice, lowestAsk, showCallLimit;
@@ -830,7 +955,7 @@ class Exchange extends React.Component {
                         }
                     }, null);
                     squeezePrice = settlementPrice / short_squeeze;
-                    showCallLimit = highestBid < lowestCallPrice && lowestCallPrice > squeezePrice;
+                    showCallLimit = lowestCallPrice > settlementPrice;
                 } else {
                     lowestAsk = asks.reduce((total, ask) => {
                         if (!total) {
@@ -841,7 +966,7 @@ class Exchange extends React.Component {
                     }, null);
 
                     squeezePrice = settlementPrice * short_squeeze;
-                    showCallLimit = lowestAsk > lowestCallPrice && lowestCallPrice < squeezePrice;
+                    showCallLimit = lowestCallPrice < settlementPrice;
                 }
             }
         }
@@ -879,6 +1004,40 @@ class Exchange extends React.Component {
         });
     }
 
+    onChangeFeeAsset(type, e) {
+        e.preventDefault();
+
+        if (type === "buy") {
+            this.setState({
+                preferCoreBuyFee: !this.state.preferCoreBuyFee
+            });
+
+            SettingsActions.changeViewSetting({
+                "preferCoreBuyFee": e.target.value === "1.3.0"
+            });
+        } else {
+            this.setState({
+                preferCoreSellFee: !this.state.preferCoreSellFee
+            });
+            
+            SettingsActions.changeViewSetting({
+                "preferCoreSellFee": e.target.value === "1.3.0"
+            });
+        }
+    }
+
+    onChangeChartHeight(increase) {
+        let newHeight = this.state.chartHeight + (increase ? 20 : -20);
+
+        this.setState({
+            chartHeight: newHeight
+        });
+
+        SettingsActions.changeViewSetting({
+            "chartHeight": newHeight
+        });
+    }
+
     render() {
         let { currentAccount, linkedAccounts, limit_orders, call_orders, totalCalls, activeMarketHistory,
             totalBids, flat_asks, flat_bids, flat_calls, invertedCalls, bids, asks, starredMarkets,
@@ -886,7 +1045,8 @@ class Exchange extends React.Component {
             marketReady, settle_orders, bucketSize } = this.props;
 
         let {buyAmount, buyPrice, buyTotal, sellAmount, sellPrice, sellTotal, leftOrderBook,
-            displayBuyPrice, displaySellPrice, buyDiff, sellDiff, indicators, indicatorSettings} = this.state;
+            displayBuyPrice, displaySellPrice, buyDiff, sellDiff, indicators, indicatorSettings,
+            width, height} = this.state;
 
         let base = null, quote = null, accountBalance = null, quoteBalance = null, baseBalance = null, coreBalance = null,
             quoteSymbol, baseSymbol, settlementPrice = null, squeezePrice = null,
@@ -922,7 +1082,7 @@ class Exchange extends React.Component {
         let quoteIsBitAsset = quoteAsset.get("bitasset_data_id") ? true : false;
         let baseIsBitAsset = baseAsset.get("bitasset_data_id") ? true : false;
 
-        let {combinedAsks, combinedBids, spread, lowestAsk, highestBid} = this._parseMarket(showCallLimit);
+        let {combinedAsks, combinedBids, spread, lowestAsk, highestBid} = this._parseMarket();
 
         // Latest price
         if (activeMarketHistory.size) {
@@ -959,7 +1119,9 @@ class Exchange extends React.Component {
         }
 
         let bucketText = function(size) {
-            if (size < 60) {
+            if (size === "all") {
+                return counterpart.translate("exchange.zoom_all");
+            } else if (size < 60) {
                 return size + "s";
             } else if (size < 3600) {
                 return (size / 60) + "m";
@@ -967,16 +1129,25 @@ class Exchange extends React.Component {
                 return (size / 3600) + "h"
             } else if (size < 604800) {
                 return (size / 86400) + "d"
-            } else if (size < 2419200) {
+            } else if (size < 2592000) {
                 return (size / 604800) + "w"
             } else {
-                return (size / 2419200) + "m"
+                return (size / 2592000) + "m"
             }
         }
 
-        let bucketOptions = buckets.map(bucket => {
-            return <div key={bucket} className={cnames("label bucket-option", {" ": bucketSize !== bucket, "active-bucket": bucketSize === bucket})} onClick={this._changeBucketSize.bind(this, bucket)}>{bucketText(bucket)}</div>
-        }).reverse();
+        let bucketOptions = buckets.filter(bucket => {
+            return bucket > 60 * 4;
+        }).map(bucket => {
+            return <div key={bucket} className={cnames("label bucket-option", {"active-bucket": bucketSize === bucket})} onClick={this._changeBucketSize.bind(this, bucket)}>{bucketText(bucket)}</div>
+        });
+
+        let oneHour = 3600;
+        let zoomPeriods = [oneHour * 6, oneHour * 48, oneHour * 48 * 2, oneHour * 24 * 7, oneHour * 24 * 14, oneHour * 24 * 30, "all"];
+        
+        let zoomOptions = zoomPeriods.map(period => {
+            return <div key={period} className={cnames("label bucket-option", {"active-bucket": this.state.currentPeriod === period})} onClick={this._changeZoomPeriod.bind(this, period)}>{bucketText(period)}</div>
+        }) ;
 
         // Market stats
         let dayChange = marketStats.get("change");
@@ -991,11 +1162,125 @@ class Exchange extends React.Component {
         let starClass = starredMarkets.has(marketID) ? "gold-star" : "grey-star";
 
         // Fees
-        let sellFee = utils.round_number(utils.get_asset_amount(this._getFee(quote), quote), quote);
-        let buyFee = utils.round_number(utils.get_asset_amount(this._getFee(base), base), base);
+        let coreAsset = ChainStore.getAsset("1.3.0");
+        if (!coreAsset) {
+            return null;
+        }
+
+        let sellFeeAsset = this.state.preferCoreSellFee ? coreAsset : quote !== coreAsset ? quote : base;
+        let sellFee = utils.round_number(utils.get_asset_amount(this._getFee(sellFeeAsset), sellFeeAsset), sellFeeAsset);
+        let sellFeeAssets = [coreAsset, quote === coreAsset ? base : quote];
+
+        let buyFeeAsset = this.state.preferCoreBuyFee ? coreAsset : base !== coreAsset ? base : quote;
+        let buyFee = utils.round_number(utils.get_asset_amount(this._getFee(buyFeeAsset), buyFeeAsset), buyFeeAsset);
+        let buyFeeAssets = [coreAsset, base === coreAsset ? quote : base];
 
         // Decimals
         let priceDecimals = Math.max(5, base ? base.get("precision") : 0);
+
+        let hasPrediction = base.getIn(["bitasset", "is_prediction_market"]) || quote.getIn(["bitasset", "is_prediction_market"]);
+
+        let description = null;
+
+        if (hasPrediction) {
+            description = quoteAsset.getIn(["options", "description"]);
+            description = assetUtils.parseDescription(description).main;
+        }
+
+        let smallScreen = false;
+        if (width < 1000) {
+            smallScreen = true;
+            leftOrderBook = false;
+        }
+
+        let buyForm = (
+            <BuySell
+                smallScreen={smallScreen}
+                style={!smallScreen && !leftOrderBook ? {minHeight: 266} : null}
+                isOpen={this.state.buySellOpen}
+                onToggleOpen={this._toggleOpenBuySell.bind(this)}
+                className={cnames("small-12 no-padding middle-content", {disabled: isNullAccount}, leftOrderBook || smallScreen ? "medium-6" : "medium-6 large-4", this.state.flipBuySell ? "order-2 sell-form" : "order-1 buy-form")}
+                type="bid"
+                amount={buyAmount}
+                price={displayBuyPrice}
+                total={buyTotal}
+                quote={quote}
+                base={base}
+                amountChange={this._buyAmountChanged.bind(this, base, quote)}
+                priceChange={this._buyPriceChanged.bind(this, base, quote)}
+                setPrice={this._currentPriceClick.bind(this, base, quote)}
+                totalChange={this._buyTotalChanged.bind(this, base, quote)}
+                balance={baseBalance}
+                onSubmit={this._createLimitOrderConfirm.bind(this, quote, base, buyAmount, buyTotal, baseBalance, coreBalance, buyFeeAsset, "buy")}
+                balancePrecision={base.get("precision")}
+                quotePrecision={quote.get("precision")}
+                totalPrecision={base.get("precision")}
+                currentPrice={lowestAsk.price_full}
+                currentPriceObject={lowestAsk.sell_price}
+                account={currentAccount.get("name")}
+                fee={buyFee}
+                feeAssets={buyFeeAssets}
+                feeAsset={buyFeeAsset}
+                onChangeFeeAsset={this.onChangeFeeAsset.bind(this, "buy")}
+                isPredictionMarket={base.getIn(["bitasset", "is_prediction_market"])}
+                onFlip={!this.state.flipBuySell ? this._flipBuySell.bind(this) : null}
+            />
+        );
+
+        let sellForm = (
+            <BuySell
+                smallScreen={smallScreen}
+                style={!smallScreen && !leftOrderBook ? {minHeight: 266} : null}
+                isOpen={this.state.buySellOpen}
+                onToggleOpen={this._toggleOpenBuySell.bind(this)}
+                className={cnames("small-12 no-padding middle-content", {disabled: isNullAccount}, leftOrderBook || smallScreen ? "medium-6" : "medium-6 large-4", this.state.flipBuySell ? "order-1 buy-form" : "order-2 sell-form")}
+                type="ask"
+                amount={sellAmount}
+                price={displaySellPrice}
+                total={sellTotal}
+                quote={quote}
+                base={base}
+                amountChange={this._sellAmountChanged.bind(this, base, quote)}
+                priceChange={this._sellPriceChanged.bind(this, base, quote)}
+                setPrice={this._currentPriceClick.bind(this, base, quote)}
+                totalChange={this._sellTotalChanged.bind(this, base, quote)}
+                balance={quoteBalance}
+                onSubmit={this._createLimitOrderConfirm.bind(this, base, quote, sellTotal, sellAmount, quoteBalance, coreBalance, sellFeeAsset, "sell")}
+                balancePrecision={quote.get("precision")}
+                quotePrecision={quote.get("precision")}
+                totalPrecision={base.get("precision")}
+                currentPrice={highestBid.price_full}
+                currentPriceObject={highestBid.sell_price}
+                account={currentAccount.get("name")}
+                fee={sellFee}
+                feeAssets={sellFeeAssets}
+                feeAsset={sellFeeAsset}
+                onChangeFeeAsset={this.onChangeFeeAsset.bind(this, "sell")}
+                isPredictionMarket={quote.getIn(["bitasset", "is_prediction_market"])}
+                onFlip={this.state.flipBuySell ? this._flipBuySell.bind(this) : null}
+            />
+        );
+
+        let orderBook = (
+            <OrderBook
+                latest={latestPrice}
+                changeClass={changeClass}
+                orders={limit_orders}
+                calls={call_orders}
+                invertedCalls={invertedCalls}
+                combinedBids={combinedBids}
+                combinedAsks={combinedAsks}
+                base={base}
+                quote={quote}
+                baseSymbol={baseSymbol}
+                quoteSymbol={quoteSymbol}
+                onClick={this._orderbookClick.bind(this, base, quote)}
+                horizontal={!leftOrderBook}
+                moveOrderBook={this._moveOrderBook.bind(this)}
+                flipOrderBook={this.props.viewSettings.get("flipOrderBook")}
+                marketReady={marketReady}
+            />
+        );
 
         return (
                 <div className="grid-block page-layout market-layout">
@@ -1005,35 +1290,29 @@ class Exchange extends React.Component {
 
                     {/* Left Column - Open Orders */}
                     {leftOrderBook ? (
-                        <div className="grid-block left-column shrink no-overflow">
-                            <OrderBook
-                                latest={latestPrice}
-                                changeClass={changeClass}
-                                orders={limit_orders}
-                                calls={call_orders}
-                                invertedCalls={invertedCalls}
-                                combinedBids={combinedBids}
-                                combinedAsks={combinedAsks}
-                                base={base}
-                                quote={quote}
-                                baseSymbol={baseSymbol}
-                                quoteSymbol={quoteSymbol}
-                                onClick={this._orderbookClick.bind(this, base, quote)}
-                                horizontal={!leftOrderBook}
-                                moveOrderBook={this._moveOrderBook.bind(this)}
-                                flipOrderBook={this.props.viewSettings.get("flipOrderBook")}
-                                marketReady={marketReady}
-                            />
+                    <div className="grid-block left-column shrink no-overflow">
+                        {orderBook}
                     </div>) : null}
 
                     {/* Center Column */}
-                    <div className={cnames("grid-block main-content vertical ps-container")} >
+                    <div className={cnames("grid-block main-content vertical no-overflow")} >
 
                         {/* Top bar with info */}
-                        <div className="grid-block no-padding shrink overflow-visible top-bar" style={{minHeight: "67px"}}>
-                            <div className="grid-block overflow-visible">
+                        <div className="grid-block no-padding shrink overflow-visible top-bar">
+                            <div className="grid-block no-overflow">
                                 <div className="grid-block shrink" style={{borderRight: "1px solid grey"}}>
-                                    <span style={{paddingRight: 0}} onClick={this._addMarket.bind(this, quoteAsset.get("symbol"), baseAsset.get("symbol"))} className="market-symbol"><Icon className={starClass} name="fi-star"/></span><Link className="market-symbol" to={`/market/${baseSymbol}_${quoteSymbol}`}><span>{`${quoteSymbol} : ${baseSymbol}`}</span></Link>
+                                    <span style={{paddingRight: 0}} onClick={this._addMarket.bind(this, quoteAsset.get("symbol"), baseAsset.get("symbol"))} className="market-symbol">
+                                        <Icon className={starClass} name="fi-star"/>
+                                    </span>
+                                    {!hasPrediction ? (
+                                        <Link className="market-symbol" to={`/market/${baseSymbol}_${quoteSymbol}`}>
+                                            <span>{`${quoteSymbol} : ${baseSymbol}`}</span>
+                                        </Link>) : (
+                                        <a className="market-symbol">
+                                            <span>{`${quoteSymbol} : ${baseSymbol}`}</span>
+                                        </a>
+                                        )}
+
                                 </div>
                                 <div className="grid-block vertical">
                                     <div className="grid-block wrap" style={{borderBottom: "1px solid grey"}}>
@@ -1081,17 +1360,24 @@ class Exchange extends React.Component {
                                     <div className="grid-block wrap no-overflow" style={{justifyContent: "space-between"}}>
                                         <ul className="market-stats stats bottom-stats">
                                             {!this.state.showDepthChart ? (
-                                                    <li className="stat" style={{minHeight: "2rem"}}>
+                                                <li className="stat">
+                                                    <span>
+                                                        <span><Translate content="exchange.zoom" />:</span>
+                                                        <span>{zoomOptions}</span>
+                                                    </span>
+                                                </li>) : null}
+                                            {!this.state.showDepthChart ? (
+                                                <li className="stat">
                                                     <span>
                                                         <span><Translate content="exchange.time" />:</span>
                                                         <span>{bucketOptions}</span>
-                                                        <span></span>
                                                     </span>
                                                 </li>) : null}
+
                                             {!this.state.showDepthChart && this.props.priceData.length ? (
                                                 <li className="stat clickable" onClick={this._onSelectIndicators.bind(this)}>
                                                     <div className="indicators">
-                                                        <Translate content="header.settings" />
+                                                        <Translate content="exchange.settings" />
                                                     </div>
                                                 </li>) : null}
                                          </ul>
@@ -1110,7 +1396,7 @@ class Exchange extends React.Component {
                                                     </div>
                                                 </li>) : null}
 
-                                                <li className="stat float-right clickable" style={{borderLeft: "1px solid grey", borderRight: "none", padding: "3px 15px"}} onClick={this._toggleCharts.bind(this)}>
+                                                <li className="stat float-right clickable" style={{borderLeft: "1px solid grey", borderRight: "none", padding: "3px 15px 0 15px"}} onClick={this._toggleCharts.bind(this)}>
                                                     <div className="indicators">
                                                        {!this.state.showDepthChart ? <Translate content="exchange.order_depth" /> : <Translate content="exchange.price_history" />}
                                                     </div>
@@ -1120,26 +1406,30 @@ class Exchange extends React.Component {
                                 </div>
                             </div>
                         </div>
-                        <div ref="center">
+                        <div className="grid-block vertical no-padding market-right-padding" id="CenterContent" ref="center">
                         {!this.state.showDepthChart ? (
                             <div className="grid-block shrink" id="market-charts" style={{marginTop: 0}}>
                                 {/* Price history chart */}
 
                                 <PriceChart
+                                    onChangeSize={this.onChangeChartHeight.bind(this)}
                                     priceData={this.props.priceData}
                                     volumeData={this.props.volumeData}
                                     base={base}
                                     quote={quote}
                                     baseSymbol={baseSymbol}
                                     quoteSymbol={quoteSymbol}
-                                    height={425}
+                                    height={this.state.height > 1100 ? this.state.chartHeight : this.state.chartHeight - 125}
                                     leftOrderBook={leftOrderBook}
                                     marketReady={marketReady}
                                     indicators={indicators}
                                     indicatorSettings={indicatorSettings}
                                     bucketSize={bucketSize}
                                     latest={latestPrice}
-                                />
+                                    verticalOrderbook={leftOrderBook}
+                                    theme={this.props.settings.get("themes")}
+                                    zoom={this.state.currentPeriod}
+                                />                                
                                 <IndicatorModal
                                     ref="indicators"
                                     indicators={indicators}
@@ -1164,46 +1454,46 @@ class Exchange extends React.Component {
                                     quote={quote}
                                     baseSymbol={baseSymbol}
                                     quoteSymbol={quoteSymbol}
-                                    height={425}
+                                    height={this.state.height > 1100 ? this.state.chartHeight : this.state.chartHeight - 125}
                                     onClick={this._depthChartClick.bind(this, base, quote)}
                                     plotLine={this.state.depthLine}
                                     settlementPrice={settlementPrice}
                                     spread={spread}
-                                    SQP={showCallLimit ? squeezePrice : null}
+                                    SQP={showCallLimit ? null : null}
                                     LCP={showCallLimit ? lowestCallPrice : null}
                                     leftOrderBook={leftOrderBook}
+                                    hasPrediction={hasPrediction}
+                                    noFrame={false}
+                                    verticalOrderbook={leftOrderBook}
+                                    theme={this.props.settings.get("themes")}
+                                />
+                            </div>)}                        
+
+                        {/* OrderBook and Market History */}
+
+                        
+                            <div className="grid-block vertical shrink buy-sell">
+                            {hasPrediction ? <div className="grid-content no-overflow" style={{lineHeight: "1.2rem", paddingTop: 10}}>{description}</div> : null}
+                            
+                            <div className="grid-block align-spaced wrap">
+                                {!leftOrderBook ? orderBook : null}
+
+                                {leftOrderBook && quote && base ? buyForm : null}
+
+                                {leftOrderBook && quote && base ? sellForm : null}
+                                
+                                <MarketHistory
+                                    className={cnames(leftOrderBook ? "hide-for-small" : "show-for-large", "no-padding no-overflow middle-content order-3 large-4")}
+                                    headerStyle={{paddingTop: 0}}
+                                    history={activeMarketHistory}
+                                    myHistory={currentAccount.get("history")}
+                                    base={base}
+                                    quote={quote}
+                                    baseSymbol={baseSymbol}
+                                    quoteSymbol={quoteSymbol}
+                                    isNullAccount={isNullAccount}
                                 />
 
-                            </div>)}
-
-                        {/* Buy/Sell forms */}
-
-                        {isNullAccount ? null : (
-                            <div className="grid-block vertical shrink buy-sell">
-                            <div className="grid-block small-vertical medium-horizontal align-spaced" style={{ flexGrow: "0" }} >
-                                {quote && base ?
-                                <BuySell
-                                    className={cnames("small-12 medium-5 no-padding", this.state.flipBuySell ? "order-3 sell-form" : "order-1 buy-form")}
-                                    type="bid"
-                                    amount={buyAmount}
-                                    price={displayBuyPrice}
-                                    total={buyTotal}
-                                    quote={quote}
-                                    base={base}
-                                    amountChange={this._buyAmountChanged.bind(this, base, quote)}
-                                    priceChange={this._buyPriceChanged.bind(this, base, quote)}
-                                    setPrice={this._currentPriceClick.bind(this, base, quote)}
-                                    totalChange={this._buyTotalChanged.bind(this, base, quote)}
-                                    balance={baseBalance}
-                                    onSubmit={this._createLimitOrderConfirm.bind(this, quote, base, buyAmount, buyTotal, baseBalance, coreBalance, "buy")}
-                                    balancePrecision={base.get("precision")}
-                                    quotePrecision={quote.get("precision")}
-                                    totalPrecision={base.get("precision")}
-                                    currentPrice={lowestAsk.price_full}
-                                    currentPriceObject={lowestAsk.sell_price}
-                                    account={currentAccount.get("name")}
-                                    fee={buyFee}
-                                /> : null}
                                 <ConfirmOrderModal
                                     type="buy"
                                     ref="buy"
@@ -1211,33 +1501,6 @@ class Exchange extends React.Component {
                                     diff={buyDiff}
                                 />
 
-                                <div onClick={this._flipBuySell.bind(this)} className="grid-block vertical align-center text-center no-padding shrink order-2" style={{cursor: "pointer"}}>
-                                    <span style={{fontSize: "2rem"}}>&#8646;</span>
-                                </div>
-                                {quote && base ?
-                                <BuySell
-                                    className={cnames("small-12 medium-5 no-padding", this.state.flipBuySell ? "order-1 buy-form" : "order-3 sell-form")}
-                                    type="ask"
-                                    amount={sellAmount}
-                                    price={displaySellPrice}
-                                    total={sellTotal}
-                                    quote={quote}
-                                    base={base}
-                                    amountChange={this._sellAmountChanged.bind(this, base, quote)}
-                                    priceChange={this._sellPriceChanged.bind(this, base, quote)}
-                                    setPrice={this._currentPriceClick.bind(this, base, quote)}
-                                    totalChange={this._sellTotalChanged.bind(this, base, quote)}
-                                    balance={quoteBalance}
-                                    onSubmit={this._createLimitOrderConfirm.bind(this, base, quote, sellTotal, sellAmount, quoteBalance, coreBalance, "sell")}
-                                    balancePrecision={quote.get("precision")}
-                                    quotePrecision={quote.get("precision")}
-                                    totalPrecision={base.get("precision")}
-                                    currentPrice={highestBid.price_full}
-                                    currentPriceObject={highestBid.sell_price}
-                                    account={currentAccount.get("name")}
-                                    fee={sellFee}
-                                    ref="sell"
-                                /> : null}
                                 <ConfirmOrderModal
                                     type="sell"
                                     ref="sell"
@@ -1245,43 +1508,52 @@ class Exchange extends React.Component {
                                     diff={sellDiff}
                                 />
                             </div>
-                        </div>)}
+                        </div>
 
-                        {!leftOrderBook ? <div className="grid-block small-12" style={{overflow: "hidden"}}>
-                            <OrderBook
-                                orders={limit_orders}
-                                calls={call_orders}
-                                invertedCalls={invertedCalls}
-                                combinedBids={combinedBids}
-                                combinedAsks={combinedAsks}
+                        <div className="grid-block no-overflow wrap shrink no-padding">
+                            {!leftOrderBook && quote && base ? buyForm : null}
+
+                            {leftOrderBook ? <ConfirmOrderModal
+                                type="buy"
+                                ref="buy"
+                                onForce={this._forceBuy.bind(this, quote, base, buyAmount, buyTotal, baseBalance, coreBalance)}
+                                diff={buyDiff}
+                            /> : null}
+
+                            {!leftOrderBook && quote && base ? sellForm : null}
+
+                            <MarketHistory
+                                className={cnames(!smallScreen && !leftOrderBook ? "medium-6 large-4" : "medium-12 large-6", {"hide-for-large" : !leftOrderBook}, "no-padding no-overflow middle-content order-3 small-12 medium-6")}
+                                headerStyle={{paddingTop: 0}}
+                                history={activeMarketHistory}
+                                myHistory={currentAccount.get("history")}
                                 base={base}
                                 quote={quote}
                                 baseSymbol={baseSymbol}
                                 quoteSymbol={quoteSymbol}
-                                onClick={this._orderbookClick.bind(this, base, quote)}
-                                horizontal={!leftOrderBook}
-                                moveOrderBook={this._moveOrderBook.bind(this)}
-                                flipOrderBook={this.props.viewSettings.get("flipOrderBook")}
+                                isNullAccount={isNullAccount}
                             />
-                    </div> : null}
 
-                        {isNullAccount ? null : (
-                            <div className="grid-block no-overflow shrink no-padding">
                             {limit_orders.size > 0 && base && quote ? (
-                                <MyOpenOrders
-                                    key="open_orders"
-                                    orders={limit_orders}
-                                    currentAccount={currentAccount.get("id")}
-                                    base={base}
-                                    quote={quote}
-                                    baseSymbol={baseSymbol}
-                                    quoteSymbol={quoteSymbol}
-                                    onCancel={this._cancelLimitOrder.bind(this)}
-                                    flipMyOrders={this.props.viewSettings.get("flipMyOrders")}
-                                />) : null}
-                        </div>)}
+                            <MyOpenOrders
+                                smallScreen={this.props.smallScreen}
+                                className={cnames({disabled: isNullAccount}, !smallScreen && !leftOrderBook ? "medium-6 large-4" : "medium-12 large-6", "small-12 no-padding align-spaced ps-container middle-content order-4")}
+                                key="open_orders"
+                                orders={limit_orders}
+                                currentAccount={currentAccount.get("id")}
+                                base={base}
+                                quote={quote}
+                                baseSymbol={baseSymbol}
+                                quoteSymbol={quoteSymbol}
+                                onCancel={this._cancelLimitOrder.bind(this)}
+                                flipMyOrders={this.props.viewSettings.get("flipMyOrders")}
+                            />) : null}
+                        </div>
+
+                    
 
                         <div className="grid-block no-overflow shrink no-padding">
+
                             {settle_orders.size > 0 && base && quote &&
                             (base.get("id") === "1.3.0" || quote.get("id") === "1.3.0") ? (
                                 <OpenSettleOrders
@@ -1303,22 +1575,9 @@ class Exchange extends React.Component {
 
 
                     {/* Right Column - Market History */}
-                    <div className="grid-block shrink right-column no-overflow vertical" style={{paddingTop: 0, paddingRight: "0.5rem"}}>
+                    <div className="grid-block shrink right-column no-overflow vertical show-for-medium" style={{paddingTop: 0, minWidth: 358, maxWidth: 400}}>
                         {/* Market History */}
-                        <div className="grid-block no-padding no-margin vertical"  style={{flex: "1 1 50vh", borderBottom: "1px solid grey"}}>
-                            <MarketHistory
-                                className="left-order-book no-padding no-overflow"
-                                headerStyle={{paddingTop: 0}}
-                                history={activeMarketHistory}
-                                myHistory={currentAccount.get("history")}
-                                base={base}
-                                quote={quote}
-                                baseSymbol={baseSymbol}
-                                quoteSymbol={quoteSymbol}
-                                isNullAccount={isNullAccount}
-                            />
-                        </div>
-                        <div className="grid-block no-padding no-margin vertical" style={{flex: "0 1 50vh"}}>
+                        <div className="grid-block no-padding no-margin vertical"  style={{borderBottom: "1px solid grey"}}>
                             <MyMarkets
                                 className="left-order-book no-padding no-overflow"
                                 headerStyle={{paddingTop: 0}}
@@ -1334,17 +1593,49 @@ class Exchange extends React.Component {
                                 current={marketID}
                             />
                         </div>
+                        <div className="grid-block no-padding no-margin vertical shrink">
+                            <DepthHighChart
+                                    orders={limit_orders}
+                                    showCallLimit={showCallLimit}
+                                    call_orders={call_orders}
+                                    flat_asks={flat_asks}
+                                    flat_bids={flat_bids}
+                                    flat_calls={ showCallLimit ? flat_calls : []}
+                                    settles={settle_orders}
+                                    invertedCalls={invertedCalls}
+                                    totalBids={totalBids}
+                                    totalCalls={showCallLimit ? totalCalls : 0}
+                                    base={base}
+                                    quote={quote}
+                                    baseSymbol={baseSymbol}
+                                    quoteSymbol={quoteSymbol}
+                                    height={200}
+                                    onClick={this._depthChartClick.bind(this, base, quote)}
+                                    plotLine={this.state.depthLine}
+                                    settlementPrice={settlementPrice}
+                                    spread={spread}
+                                    SQP={showCallLimit ? squeezePrice : null}
+                                    LCP={showCallLimit ? lowestCallPrice : null}
+                                    leftOrderBook={leftOrderBook}
+                                    hasPrediction={hasPrediction}
+                                    noText={true}
+                                    theme={this.props.settings.get("themes")}
+                                />                                
+                        </div>
                     </div>
-                    {!isNullAccount && quoteIsBitAsset ?
+
+                    {!isNullAccount && quoteIsBitAsset  ?
                         <BorrowModal
                             ref="borrowQuote"
                             quote_asset={quoteAsset.get("id")}
+                            backing_asset={quoteAsset.getIn(["bitasset", "options", "short_backing_asset"])}
                             account={currentAccount}
                          /> : null}
-                    {!isNullAccount && baseIsBitAsset ?
+                    {!isNullAccount && baseIsBitAsset ? 
                         <BorrowModal
                             ref="borrowBase"
                             quote_asset={baseAsset.get("id")}
+                            backing_asset={baseAsset.getIn(["bitasset", "options", "short_backing_asset"])}
                             account={currentAccount}
                         /> : null}
                 {/* End of Second Vertical Block */}
