@@ -20,6 +20,235 @@ import {debounce} from "lodash";
 
 let lastLookup = new Date();
 
+class MarketGroup extends React.Component {
+
+    static defaultProps = {
+        maxRows: 20
+    };
+
+    constructor(props) {
+        super();
+        this.state = this._getInitialState(props);
+    }
+
+    _getInitialState(props) {
+        let open = props.forceOpen ? true : props.viewSettings.get(`myMarketsBase_${props.index}`);
+        return {
+            open: open !== undefined ? open : true,
+            inverseSort: props.viewSettings.get("myMarketsInvert", true),
+            sortBy: props.viewSettings.get("myMarketsSort", "volume")
+        };
+    }
+
+    componentWillReceiveProps(nextProps) {
+        if (nextProps.forceOpen !== this.props.forceOpen) {
+            this.setState(this._getInitialState(nextProps));
+        }
+    }
+
+    shouldComponentUpdate(nextProps, nextState) {
+        if (!nextProps.markets || !this.props.markets) {
+            return true;
+        }
+        return (
+            !utils.are_equal_shallow(nextState, this.state) ||
+            !utils.are_equal_shallow(nextProps.markets, this.props.markets) ||
+            nextProps.starredMarkets !== this.props.starredMarkets ||
+            nextProps.marketStats !== this.props.marketStats
+        );
+    }
+
+    _inverseSort() {
+        SettingsActions.changeViewSetting({
+            myMarketsInvert: !this.state.myMarketsInvert
+        });
+        this.setState({
+            inverseSort: !this.state.inverseSort
+        });
+    }
+
+    _changeSort(type) {
+        if (type !== this.state.sortBy) {
+            SettingsActions.changeViewSetting({
+                myMarketsSort: type
+            });
+            this.setState({
+                sortBy: type
+            });
+        } else {
+            this._inverseSort();
+        }
+    }
+
+    // _onSelectBase(e) {
+    //     e.preventDefault();
+    //     e.stopPropagation();
+
+    //     SettingsActions.changeBase(this.props.index, e.target.value);
+    // }
+
+    _onToggle(e) {
+        if (!this.props.forceOpen) {
+            let open = !this.state.open;
+            this.setState({
+                open: open
+            });
+
+            let setting = {};
+            setting[`myMarketsBase_${this.props.index}`] = open;
+            SettingsActions.changeViewSetting(setting);
+        }
+    }
+
+    render() {
+        let {columns, markets, base, marketStats, starredMarkets,
+            current, maxRows, baseOptions} = this.props;
+        let {sortBy, inverseSort, open} = this.state;
+
+        if (!markets || !markets.length) {
+            return null;
+        }
+
+        let headers = columns.map(header => {
+            switch (header.name) {
+                case "market":
+                    return <th key={header.name} className="clickable" onClick={this._changeSort.bind(this, "name")}><Translate content="exchange.market" /></th>;
+
+                case "vol":
+                    return <th key={header.name} className="clickable" onClick={this._changeSort.bind(this, "volume")}style={{textAlign: "right"}}><Translate content="exchange.vol_short" /></th>;
+
+                case "price":
+                    return <th key={header.name} style={{textAlign: "right"}}><Translate content="exchange.price" /></th>;
+
+                case "quoteSupply":
+                    return <th key={header.name}><Translate content="exchange.quote_supply" /></th>;
+
+                case "baseSupply":
+                    return <th key={header.name}><Translate content="exchange.base_supply" /></th>;
+
+                case "change":
+                    return <th key={header.name} className="clickable" onClick={this._changeSort.bind(this, "change")} style={{textAlign: "right"}}><Translate content="exchange.change" /></th>;
+
+                default:
+                    return <th key={header.name}></th>;
+            }
+        });
+
+        let index = 0;
+
+        let marketRows = markets
+            .map(market => {
+
+                return (
+                    <MarketRow
+                        key={market.id}
+                        name={base === "others" ? market.quote + ":" + market.base : market.quote}
+                        quote={market.quote}
+                        base={market.base}
+                        columns={columns}
+                        leftAlign={true}
+                        compact={true}
+                        noSymbols={true}
+                        stats={marketStats.get(market.id)}
+                        starred={starredMarkets.has(market.id)}
+                        current={current === market.id}
+                    />
+                );                   
+            }).filter(a => {
+                return a !== null;
+            }).sort((a, b) => {
+                let a_symbols = a.key.split("_");
+                let b_symbols = b.key.split("_");
+                let aStats = marketStats.get(a_symbols[0] + "_" + a_symbols[1]);
+                let bStats = marketStats.get(b_symbols[0] + "_" + b_symbols[1]);
+
+                switch (sortBy) {
+
+                    case "name":
+                        if (a_symbols[0] > b_symbols[0]) {
+                            return inverseSort ? -1 : 1;
+                        } else if (a_symbols[0] < b_symbols[0]) {
+                            return inverseSort ? 1 : -1;
+                        } else {
+                            if (a_symbols[1] > b_symbols[1]) {
+                                return inverseSort ? -1 : 1;
+                            } else if (a_symbols[1] < b_symbols[1]) {
+                                return inverseSort ? 1 : -1
+                            } else {
+                                return 0;
+                            }
+                        }
+
+                    case "volume":
+                        if (aStats && bStats) {
+                            if (inverseSort) {
+                                return bStats.volumeBase - aStats.volumeBase;
+                            } else {
+                                return aStats.volumeBase - bStats.volumeBase;
+                            }
+                        } else {
+                            return 0;
+                        }
+
+                    case "change":
+                        if (aStats && bStats) {
+                            if (inverseSort) {
+                                return bStats.change - aStats.change;
+                            } else {
+                                return aStats.change - bStats.change;
+                            }
+                        } else {
+                            return 0;
+                        }
+                }
+            }).filter(() => {
+                ++index;
+                return index < maxRows;
+            });
+
+        let options = baseOptions.map((option, index) => {
+            return <option key={index} value={option}>{option}</option>;
+        });
+
+        let caret = open ? <span>&#9660;</span> : <span>&#9650;</span>;
+
+        return (
+            <div style={{paddingRight: 10}}>
+            <div
+                id={base}
+                onClick={this._onToggle.bind(this)}
+                className="mymarkets-group-header"
+            >
+                <div style={{zIndex: 10}}>
+                    {base !== "others" && this.props.allowChange ?
+                    (<span>
+                        <select
+                            onChange={this._onSelectBase.bind(this)}
+                            className="mymarkets-base-select"
+                            value={base}
+                        >
+                            {options}
+                        </select>
+                         ({marketRows.length})
+                    </span>) :
+                    base !== "others" ? <span>{base} ({marketRows.length})</span> : <span><Translate content="exchange.others" /> ({marketRows.length})</span>}
+                    {!this.props.forceOpen ? <div className="float-right">{caret}</div> : null}
+                </div>
+            </div>
+                {open ? (
+                <table className="table table-hover text-right">
+                    <thead>
+                        <tr>{headers}</tr>
+                    </thead>
+                    {marketRows && marketRows.length ?
+                        <tbody>{marketRows}</tbody> : null
+                    }
+                </table>) : null}
+            </div>
+        );
+    }
+}
+
 @BindToChainState()
 class MyMarkets extends React.Component {
 
@@ -29,7 +258,8 @@ class MyMarkets extends React.Component {
 
     static defaultProps = {
         activeTab: "starred",
-        core: "1.3.0"
+        core: "1.3.0",
+        setMinWidth: false
     };
 
     static contextTypes = {
@@ -39,20 +269,22 @@ class MyMarkets extends React.Component {
     constructor(props) {
         super();
 
-        let inputValue = props.viewSettings.get("marketLookupInput") || null;
+        let inputValue = props.viewSettings.get("marketLookupInput", null);
         let symbols = inputValue ? inputValue.split(":") : [null];
         let quote = symbols[0];
         let base = symbols.length === 2 ? symbols[1] : null;
 
         this.state = {
-            inverseSort: props.viewSettings.get("myMarketsInvert") || true,
-            sortBy: props.viewSettings.get("myMarketsSort") || "volume",
-            activeTab: props.viewSettings.get("favMarketTab") || "starred",
+            inverseSort: props.viewSettings.get("myMarketsInvert", true),
+            sortBy: props.viewSettings.get("myMarketsSort", "volume"),
+            activeTab: props.viewSettings.get("favMarketTab", "starred"),
             lookupQuote: quote,
             lookupBase: base,
-            inputValue: inputValue
+            inputValue: inputValue,
+            minWidth: "100%"
         };
 
+        this._setMinWidth = this._setMinWidth.bind(this);
         this.getAssetList = _.debounce(AssetActions.getAssetList, 150);
     }
 
@@ -68,7 +300,10 @@ class MyMarkets extends React.Component {
             nextState.activeTab !== this.state.activeTab ||
             nextState.lookupQuote !== this.state.lookupQuote ||
             nextState.lookupBase !== this.state.lookupBase ||
-            nextProps.current !== this.props.current
+            nextProps.current !== this.props.current ||
+            nextProps.minWidth !== this.props.minWidth ||
+            nextProps.listHeight !== this.props.listHeight ||
+            nextProps.preferredBases !== this.props.preferredBases
         );
     }
 
@@ -81,18 +316,37 @@ class MyMarkets extends React.Component {
         if (this.state.activeTab === "all" && this.state.inputValue) {
             this._lookupAssets({target: {value: this.state.inputValue}}, true);
         }
+
+    }
+
+    componentWillMount() {
+        if (this.props.setMinWidth) {
+            window.addEventListener("resize", this._setMinWidth, false);
+        }
+    }
+
+    componetWillUnmount() {
+        if (this.props.setMinWidth) {
+            window.removeEventListener("resize", this._setMinWidth, false);
+        }
     }
 
     _setMinWidth() {
-        if (this.refs.favorites && this.props.activeTab === "starred") {
-            this.setState({
-                minWidth: this.refs.favorites.offsetWidth
-            });
+        if (this.props.setMinWidth && this.refs.favorites && this.props.activeTab === "starred") {
+            
+            if (this.state.minWidth !== this.refs.favorites.offsetWidth) {
+                this.setState({
+                    minWidth: this.refs.favorites.offsetWidth
+                });
+            }
         }
     }
+
     componentDidUpdate() {
-        let historyContainer = ReactDOM.findDOMNode(this.refs.favorites);
-        Ps.update(historyContainer);
+        if (this.refs.favorites) {
+            let historyContainer = ReactDOM.findDOMNode(this.refs.favorites);
+            Ps.update(historyContainer);
+        }
     }
 
     _inverseSort() {
@@ -168,21 +422,24 @@ class MyMarkets extends React.Component {
     }
 
     render() {
-        let {starredMarkets, marketStats, columns, searchAssets, core, current} = this.props;
+        let {starredMarkets, marketStats, columns, searchAssets,
+            preferredBases, core, current, viewSettings, listHeight, baseOptions} = this.props;
         let {inverseSort, activeTab, sortBy, lookupQuote, lookupBase} = this.state;
-        let marketRows = <tr></tr>;
+        let otherMarkets = <tr></tr>;
 
         let coreSymbol = core.get("symbol");
         // Add some default base options
-        let defaultBases = [coreSymbol, "BTC", "CNY", "USD"];
-        let baseOptions = [
+        // let preferredBases = [coreSymbol, "BTC", "USD", "CNY"];
+        let baseGroups = {};
+
+        let bases = [
             // coreSymbol, "BTC", "CNY", "USD"
         ];
 
         searchAssets
         .filter(a => {
             // Always keep core asset as an option
-            // if (defaultBases.indexOf(a.symbol) === 0) {
+            // if (preferredBases.indexOf(a.symbol) === 0) {
             //     return true;
             // }
             if (lookupBase && lookupBase.length) {
@@ -193,21 +450,26 @@ class MyMarkets extends React.Component {
         .forEach(asset => {
             if (lookupBase && lookupBase.length) {
                 if (asset.symbol.indexOf(lookupBase) === 0) {
-                    baseOptions.push(asset.symbol);
+                    bases.push(asset.symbol);
                 }
-            } else if (defaultBases.indexOf(asset.symbol) < 0 ) {
+            } else if (preferredBases.includes(asset.symbol)) {
                 if (asset.symbol.length >= lookupQuote.length && asset.symbol.length < lookupQuote.length + 3) {
-                    baseOptions.push(asset.symbol);
+                    bases.push(asset.symbol);
                 }
             }
         });
 
-        baseOptions = baseOptions.concat(defaultBases.filter(a => {if (!lookupBase || !lookupBase.length) {return true}; return a.indexOf(lookupBase) === 0;}));
+        bases = bases.concat(preferredBases.filter(a => {
+            if (!lookupBase || !lookupBase.length) {
+                return true
+            }
+            return a.indexOf(lookupBase) === 0;
+        }).toArray());
 
-        baseOptions = baseOptions
+        bases = bases
         .filter(base => {
             // Always keep core asset as an option
-            // if (defaultBases.indexOf(base) !== -1) {
+            // if (preferredBases.indexOf(base) !== -1) {
             //     return true;
             // }
             if (lookupBase && lookupBase.length > 1) {
@@ -228,7 +490,7 @@ class MyMarkets extends React.Component {
                 );
             })
             .forEach(asset => {
-                baseOptions.forEach(base => {
+                bases.forEach(base => {
                     let marketID = asset.symbol + "_" + base;
 
                     if (base !== asset.symbol) {
@@ -245,14 +507,13 @@ class MyMarkets extends React.Component {
                 return a[1].quote === lookupQuote;
             }
             return true;
-        })
+        });
 
         allMarkets = Immutable.Map(allMarkets);
-
         let activeMarkets = activeTab === "starred" ? starredMarkets : allMarkets;
 
         if (activeMarkets.size > 0) {
-            marketRows = activeMarkets
+            otherMarkets = activeMarkets
             .filter(a => {
                 if (activeTab === "all") {
                     if (lookupQuote.length < 2) {return false; }
@@ -263,70 +524,32 @@ class MyMarkets extends React.Component {
             })
             .map(market => {
                 let marketID = market.quote + "_" + market.base;
-                    return (
-                        <MarketRow
-                            key={marketID}
-                            quote={market.quote}
-                            base={market.base}
-                            columns={columns}
-                            leftAlign={true}
-                            compact={true}
-                            noSymbols={true}
-                            stats={marketStats.get(marketID)}
-                            starred={starredMarkets.has(marketID)}
-                            current={current === marketID}
-                        />
+                if (preferredBases.includes(market.base)) {
+                    if (!baseGroups[market.base]) {
+                        baseGroups[market.base] = [];
+                    }
+                    baseGroups[market.base].push(
+                        {
+                            id: marketID,
+                            quote: market.quote,
+                            base: market.base
+                        }
                     );
+                    return null;
+                } else {
+                    return (
+                        {
+                            id: marketID,
+                            quote: market.quote,
+                            base: market.base
+                        }
+                    ); 
+                }
+                   
             }).filter(a => {
                 return a !== null;
-            }).sort((a, b) => {
-                let a_symbols = a.key.split("_");
-                let b_symbols = b.key.split("_");
-                let aStats = marketStats.get(a_symbols[0] + "_" + a_symbols[1]);
-                let bStats = marketStats.get(b_symbols[0] + "_" + b_symbols[1]);
-
-                switch (sortBy) {
-
-                    case "name":
-                        if (a_symbols[0] > b_symbols[0]) {
-                            return inverseSort ? -1 : 1;
-                        } else if (a_symbols[0] < b_symbols[0]) {
-                            return inverseSort ? 1 : -1;
-                        } else {
-                            if (a_symbols[1] > b_symbols[1]) {
-                                return inverseSort ? -1 : 1;
-                            } else if (a_symbols[1] < b_symbols[1]) {
-                                return inverseSort ? 1 : -1
-                            } else {
-                                return 0;
-                            }
-                        }
-
-                    case "volume":
-                        if (aStats && bStats) {
-                            if (inverseSort) {
-                                return bStats.volumeQuote - aStats.volumeQuote;
-                            } else {
-                                return aStats.volumeQuote - bStats.volumeQuote;
-                            }
-                        } else {
-                            return 0;
-                        }
-
-                    case "change":
-                        if (aStats && bStats) {
-                            if (inverseSort) {
-                                return bStats.change - aStats.change;
-                            } else {
-                                return aStats.change - bStats.change;
-                            }
-                        } else {
-                            return 0;
-                        }
-                }
-
             })
-            .take(activeTab === "starred" ? 100 : 30)
+            .take(activeTab === "starred" ? 100 : 20)
             .toArray();
         }
 
@@ -334,34 +557,25 @@ class MyMarkets extends React.Component {
         let starClass = cnames(hc, {inactive: activeTab === "all"});
         let allClass = cnames(hc, {inactive: activeTab === "starred"});
 
-        let headers = columns.map(header => {
-            switch (header.name) {
-                case "market":
-                    return <th key={header.name} className="clickable" onClick={this._changeSort.bind(this, "name")}><Translate content="exchange.market" /></th>;
+        let listStyle = {
+            minWidth: this.state.minWidth,
+            paddingTop: 8
+        };
+        if (listHeight) {
+            listStyle.height = listHeight;
+        }
 
-                case "vol":
-                    return <th key={header.name} className="clickable" onClick={this._changeSort.bind(this, "volume")}style={{textAlign: "right"}}><Translate content="exchange.vol_short" /></th>;
-
-                case "price":
-                    return <th key={header.name} style={{textAlign: "right"}}><Translate content="exchange.price" /></th>;
-
-                case "quoteSupply":
-                    return <th key={header.name}><Translate content="exchange.quote_supply" /></th>;
-
-                case "baseSupply":
-                    return <th key={header.name}><Translate content="exchange.base_supply" /></th>;
-
-                case "change":
-                    return <th key={header.name} className="clickable" onClick={this._changeSort.bind(this, "change")} style={{textAlign: "right"}}><Translate content="exchange.change" /></th>;
-
-                default:
-                    return <th key={header.name}></th>;
-            }
+        // Filter out already picked baseOptions
+        baseOptions = baseOptions.filter(a => {
+            return !preferredBases.includes(a);
         });
 
         return (
             <div className={this.props.className} style={this.props.style}>
-                <div style={this.props.headerStyle} className="grid-block shrink left-orderbook-header bottom-header">
+                <div
+                    style={this.props.headerStyle}
+                    className="grid-block shrink left-orderbook-header bottom-header"
+                >
                     <div ref="myMarkets" className={starClass} onClick={this._changeTab.bind(this, "starred")}>
                         <Translate content="exchange.market_name" />
                     </div>
@@ -375,15 +589,43 @@ class MyMarkets extends React.Component {
                         {this.props.controls ? <div style={{paddingBottom: "0.5rem"}}>{this.props.controls}</div> : null}
                         {activeTab === "all" ? <input type="text" value={this.state.inputValue} onChange={this._lookupAssets.bind(this)} placeholder="SYMBOL:SYMBOL" /> : null}
                     </div> ) : null}
-                <div style={{minWidth: this.state.minWidth}} className="table-container grid-content mymarkets-list" ref="favorites">
-                    <table className="table table-hover text-right market-right-padding">
-                        <thead>
-                            <tr>{headers}</tr>
-                        </thead>
-                        {marketRows && marketRows.length ?
-                            <tbody>{marketRows}</tbody> : null
-                        }
-                    </table>
+
+                <div
+                    style={listStyle}
+                    className="table-container grid-block vertical mymarkets-list"
+                    ref="favorites"
+                >
+                    {preferredBases.map((base, index) => {
+                        return <MarketGroup
+                            index={index}
+                            allowChange={false}
+                            key={base}
+                            current={current}
+                            starredMarkets={starredMarkets}
+                            baseOptions={baseOptions}
+                            marketStats={marketStats}
+                            viewSettings={viewSettings}
+                            columns={columns}
+                            markets={baseGroups[base]}
+                            base={base}
+                            maxRows={activeTab === "starred" ? 20 : 10}
+                            forceOpen={activeTab === "all"}
+                        />
+                    })}
+
+                    <MarketGroup
+                        index={preferredBases.size}
+                        current={current}
+                        starredMarkets={starredMarkets}
+                        baseOptions={baseOptions}
+                        marketStats={marketStats}
+                        viewSettings={viewSettings}
+                        columns={columns}
+                        markets={otherMarkets}
+                        base="others"
+                        maxRows={activeTab === "starred" ? 20 : 10}
+                        forceOpen={activeTab === "all"}
+                    />
                 </div>
             </div>
         );
@@ -400,6 +642,8 @@ class MyMarketsWrapper extends React.Component {
         return {
             starredMarkets: SettingsStore.getState().starredMarkets,
             viewSettings: SettingsStore.getState().viewSettings,
+            preferredBases: SettingsStore.getState().preferredBases,
+            baseOptions: SettingsStore.getState().baseOptions,
             marketStats: MarketsStore.getState().allMarketStats,
             searchAssets: AssetStore.getState().assets
         }
