@@ -284,9 +284,8 @@ class CommitteeMembers extends React.Component {
                         </div>
 
                     </div>
-                    <div className="grid-block">
-                        <div className="grid-content">
-                            <div className="grid-block small-12 medium-6">
+                    <div className="grid-block vertical">
+                            <div className="grid-block vertical small-12 medium-6">
                                 <Translate component="h3" content="markets.filter" />
                                 <input type="text" value={this.state.filterCommitteeMember} onChange={this._onFilter.bind(this)} />
                             </div>
@@ -297,7 +296,6 @@ class CommitteeMembers extends React.Component {
                                 cardView={this.state.cardView}
                             />
                         </div>
-                    </div>
                 </div>
             </div>
         );
