@@ -29,7 +29,7 @@ export default class BrowserSupportModal extends React.Component {
 
                     <div className="button-group no-overflow" style={{paddingTop: "0"}}>
                         <Trigger close="browser_modal">
-                            <a href onClick={this._openLink} className="button"><Translate content="init_error.understand" /></a>
+                            <div href onClick={this._openLink} className="button"><Translate content="init_error.understand" /></div>
                         </Trigger>
                     </div>
 
