@@ -116,11 +116,11 @@ class WalletUnlockModal extends React.Component {
                 </Trigger>
                 
                 <h3><Translate content="header.unlock" /> {unlock_what}</h3>
-                
+
                 <form onSubmit={this.onPasswordEnter} noValidate autoComplete="off">
                     
                     <AuthInput ref="auth_input" auth={this.props.auth} />
-                    
+
                     <div className="button-group">
                         <button 
                             className={cname("button", {disabled: this.state.working || !this.props.auth.valid}) }
